ROSEx

ROSEx is a differentiable modular modeling framework for use with very long baseline interferometry. The goal is to allow the user to easily combine and modify a set of primitive models to construct complicated source structures. These primitives themselves do not have to be simple. ROSE itself does not Bayesian inference or optimization itself. Instead it creates all the components needed, i.e. a image/visibility model, some simple likelihoods, and telescope corruption effects (still to be implemented).

To use perform inferences on data you can then hook into the vast array of different modeling and optimization packages in Julia. There are some small examples packages defining these interface such as ROSESoss.jl which combines ROSE with Soss a probabilistic programming language. Other interfaces to e.g. Turing, BAT are planned.

Requirements

The minimum Julia version we require is 1.6, which is the current LTS release. In the future we may increase this as Julia advances.

ROSEx

Requirements

Getting Started

ROSEx.ROSEx — Module

ROSEx

Radio Observation Sampling Exploration

ROSEx.AbstractCache — Type

abstract type AbstractCache

This defines an abstract cache that can be used to hold or precompute some computations.

ROSEx.AbstractModifier — Type

abstract type AbstractModifier{M<:ROSEx.AbstractModel} <: ROSEx.AbstractModel

Abstract type for image modifiers. These are some model wrappers that can transform any model using simple Fourier transform properties. To see the implemented modifier

ROSEx.AddModel — Type

struct AddModel{T1, T2} <: ROSEx.CompositeModel{T1, T2}

Adds two models together to create composite models. Note that I may change this in the future so make it easier on the compiler, i.e. make the composite model a fancy model vector and heap allocate stuff. This should help when combining multiple models together.

ROSEx.ArrayConfiguration — Type

abstract type ArrayConfiguration

This defined the abstract type for an array configuration. Namely, baseline times, SEFD's, bandwidth, observation frequencies, etc.

ROSEx.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

ROSEx.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ROSEx.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

intensity_point if model intensity is IsAnalytic

intensitymap! if model intensity is NotAnalytic

flux

ROSEx.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: ROSEx.GeometricModel

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using ROSEx.jl's model composition functionality.

ROSEx.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: ROSEx.CompositeModel{M1, M2}

Smooths a model with a Gaussian kernel with standard deviation of σ.

Notes

m = smoothed(Disk(), 5.0)
intensity(m, 2.0, 2.0; fov=10.0, npix=128)

will compute the intensity of the smoothed disk with using an interpolation with 128 pixel nodes and a total fov of 10 in x and y direction.

This needs to be improved

ROSEx.DeltaPulse — Type

struct DeltaPulse{T} <: ROSEx.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

ROSEx.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ROSEx.EHTClosurePhaseDatum — Type

struct EHTClosurePhaseDatum{T<:Number} <: ROSEx.ClosureProducts{T<:Number}

A Datum for a single closure phase. Note in the future this may get replaced with the fully covariant formalism from Blackburn et al. (2020).

ROSEx.ExtendedRing — Type

struct ExtendedRing{F} <: ROSEx.GeometricModel

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

Note

@e mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

radius

radius of peak emission

shape

shape of the radial distribution

ROSEx.FFT — Type

struct FFT <: ROSEx.FourierTransform

Fourier transform type that specifies we will use the FFTW package to compute the Fourier transform.

Fields

padfac

The amount to pad the image by. Note we actually round up to the nearest factor of 2, but this will be improved in the future to use small primes

ROSEx.FFTCache — Type

struct FFTCache{P, I} <: ROSEx.AbstractCache

The cache used when the FFT algorithm is used to compute visibilties. This is an internal type and is not part of the public API

Fields

plan

FFTW Plan

sitp

FFT interpolator function

ROSEx.FourierTransform — Type

abstract type FourierTransform

Abstract type that specified which fourier transform to use

ROSEx.Gaussian — Type

struct Gaussian{T} <: ROSEx.GeometricModel

Gaussian geometrical model. This is a Gaussian with unit flux and standard deviation.

Notes

To change the Gaussian flux, and shape please use the modifier functions

ROSEx.GeometricModel — Type

abstract type GeometricModel <: ROSEx.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain.

ROSEx.IsAnalytic — Type

struct IsAnalytic <: ROSEx.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ROSEx.MRing — Type

struct MRing{T, N} <: ROSEx.GeometricModel

m-ring geometric model. This corresponds to a delta ring with a fourier expansion in θ. The m in m-ring refers to the order of the Fourier expansion.

ROSEx.NoCache — Type

struct NoCache <: ROSEx.AbstractCache

No cache is used. This is typically used when the model is analytic in the Fourier domain.

ROSEx.NotAnalytic — Type

struct NotAnalytic <: ROSEx.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ROSEx.PolarizedModel — Type

struct PolarizedModel{I, Q, U, V} <: ROSEx.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

ROSEx.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into ROSEx more easily.

ROSEx.Pulse — Type

Pulse Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

ROSEx.RImage — Type

$(TYPEDEF)

An image model given by a set of coefficients and a kernel response or basis function. This corresponds to a continous image defined by a finite set of points. The defined intensity is given by

\[ I(x,y) = \sum_{ij} c_{ij}κ(x-x_i)κ(y-y_i).\]

An important thing to note is that the $c_{ij}$ do not represent pixel intensities, i.e. the κ doesn't have to be an interpolating kernel.

Example

samples = rand(10,10)
model = RImage(samples, BSplineKernel{3})

Notes

This is defined in terms of pixel response, so the image size is 1μas. To resize the image use the scale function like with other models.

Fields
FIELDS

ROSEx.RenormalizedModel — Type

struct RenormalizedModel{M<:ROSEx.AbstractModel, T} <: ROSEx.AbstractModifier{M<:ROSEx.AbstractModel}

Renormalizes the flux of the model to the new value flux. We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

ROSEx.RotatedModel — Type

struct RotatedModel{M<:ROSEx.AbstractModel, T} <: ROSEx.AbstractModifier{M<:ROSEx.AbstractModel}

Type for the rotated model. This is more fine grained constrol of rotated model. In most use cases the end-user should be using the rotate method e.g.

rotate(model, ξ)

ROSEx.ShiftedModel — Type

struct ShiftedModel{T, M<:ROSEx.AbstractModel} <: ROSEx.AbstractModifier{M<:ROSEx.AbstractModel}

Shifts the model by Δx units in the x-direction and Δy units in the y-direction.

ROSEx.SqExpPulse — Type

struct SqExpPulse{T} <: ROSEx.Pulse

Normalized square exponential kernel, i.e. a Gaussian. Note the smoothness is modfied with ϵ which is the inverse variance in units of 1/pixels².

ROSEx.StretchedModel — Type

struct StretchedModel{M<:ROSEx.AbstractModel, T} <: ROSEx.AbstractModifier{M<:ROSEx.AbstractModel}

Stretched the model in the x and y directions, i.e. the new intensity is

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

Base.size — Method

size(model)

return the size of the coefficient matrix for model.

ROSEx.amplitude — Method

amplitude(model, u, v)

Computes the visibility amplitude of model m at u,v positions u,v

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.amplitudes — Method

amplitudes(m, u, v)

Computes the amplitudes of the model m at the u,v positions u, v.

Note this is done lazily so the visibility is only computed when accessed.

ROSEx.basemodel — Method

basemodel(model)

Returns the base model from a modified model.

ROSEx.bispectra — Method

bispectra(m, u1, v1, u2, v2, u3, v3)

Computes the bispectra of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the bispectra is only computed when accessed.

ROSEx.bispectrum — Method

bispectrum(model, u1, v1, u2, v2, u3, v3)

Computes the complex bispectrum of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.closure_phase — Method

closure_phase(model, u1, v1, u2, v2, u3, v3)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.closure_phase — Method

closure_phase(D1, D2, D3)

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

ROSEx.closure_phases — Method

closure_phases(m, u1, v1, u2, v2, u3, v3)

Computes the closure phases of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the closure_phases is only computed when accessed.

ROSEx.coherencymatrix — Method

coherencymatrix(pimg, u, v)

Computes the coherency matrix of the polarized model.

ROSEx.components — Method

components(m)

Returns the components for a composite model. This will return a Tuple with all the models you have constructed.

ROSEx.create_cache — Method

create_cache(alg, model, img)

Creates the model cache given for the algorithm alg using the model and a image cache image

ROSEx.evpa — Method

evpa(pimg, u, v)

electric vector position angle or EVPA of the polarized model

ROSEx.extract_cphase — Method

extract_cphase(obs)

Extracts the closure phases from an ehtim observation object

Returns an EHTObservation with closure phases datums

ROSEx.extract_lcamp — Method

extract_lcamp(obs)

Extracts the log-closure amp. from an ehtim observation object

Returns an EHTObservation with closure amp. datums

ROSEx.extract_vis — Method

extract_vis(obs)

Extracts the complex visibilities from an ehtim observation object

Returns an EHTObservation with complex visibility data

ROSEx.flux — Method

Computes the flux of a intensity map

ROSEx.imanalytic — Method

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ROSEx.intensity_point — Function

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ROSEx.intensitymap — Function

Computes the intensity map of model. This version requires additional information to construct the grid.

Example

m = Gaussian()
# field of view
fovx, fovy = 5.0
fovy = 5.0
# number of pixels
nx, ny = 128

img = intensitymap(m, fovx, fovy, nx, ny; pulse=DeltaPulse())

ROSEx.intensitymap! — Function

Computes the intensity map of model by modifying the input IntensityMap object

ROSEx.intensitymap! — Method

intensitymap!(pimg, pmodel)

Finds the polarized intensity map of the polarized model pmodel.

ROSEx.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive ROSEx model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of ROSEx they need to specify if it is primitive

struct MyPrimitiveModel end
ROSEx.isprimitive(::Type{MyModel}) = ROSEx.IsPrimitive()

ROSEx.load_ehtim — Method

load_ehtim()

Loads the eht-imaging library and stores it in the ehtim variable.

Notes

This will fail if ehtim isn't installed in the python installation that PyCall references.

ROSEx.load_tpy — Method

load_tpy(file)

Load a ThemisPy style ascii EHT observation file.

ROSEx.logclosure_amplitude — Method

logclosure_amplitude(model, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v3 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.logclosure_amplitudes — Method

logclosure_amplitudes(m, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log closure amplitudes of the model m at the quadrangles (u1,v1), (u2,v2), (u3,v3), (u4, v4).

Note this is done lazily so the log closure amplitude is only computed when accessed.

ROSEx.modelimage — Method

modelimage(m)

Construct a ModelImage where just the model m is specified

Notes

Currently this is only defined for analytic models. In the future this will guess a reasonable image to use.

ROSEx.modelimage — Method

modelimage(model, image; alg)
modelimage(m)

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg

Notes

For analytic models this is a no-op and just return the model. For non-analytic models this wraps the model in a object with an image and precomputes the fourier transform using alg.

ROSEx.m̆ — Method

m̆(pimg, u, v)

Computes the fractional linear polarization in the visibility domain

\[ \breve{m} = \frac{Q + iU}{I}\]

ROSEx.rad2μas — Method

rad2μas(x)

Converts a number from radians to μas

ROSEx.renormed — Method

renormed(model, flux)

Renormalizes the model m to have total flux flux.

ROSEx.rotated — Method

rotated(model, ξ)

Rotates the model by an amount ξ in radians.

ROSEx.shifted — Method

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy.

ROSEx.stretched — Method

stretched(model, α, β)

Stretches the model m according to the formula

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

ROSEx.visanalytic — Method

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ROSEx.visibilities — Method

visibilities(m, u, v)

Computes the visibilities of the model m at the u,v positions u, v.

Note this is done lazily so the visibility is only computed when accessed.

ROSEx.visibility — Function

Computes the complex visibility of model m at u,v positions u,v

If you want to compute the visibilities at a large number of positions consider using the visibilities function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.visibility — Method

visibility(pimg, u, v)

Computes the visibility in the stokes basis of the polarized model

ROSEx.visibility — Method

visibility(mimg, u, v)

Computes the complex visibility of model m at u,v positions u,v

If you want to compute the visibilities at a large number of positions consider using the visibilities function which uses MappedArrays to compute a lazy, no allocation vector.

ROSEx.visibility_point — Function

Function that computes the pointwise visibility if the model has the trait in the fourier domain IsAnalytic(). Otherwise it will use the FFTW fallback.

ROSEx.μas2rad — Method

μas2rad(x)

Converts a number from μas to rad