Comrade API

Contents

Comrade API

Contents

Index

Model Definitions

Data Types

eht-imaging interface

Bayesian Tools

Misc

Index

Comrade.AbstractModifier

Comrade.AddModel

Comrade.AmpNormal

Comrade.ArrayBaselineDatum

Comrade.CPVonMises

Comrade.ClosureConfig

Comrade.ComplexNormal

Comrade.CompositeModel

Comrade.ConcordanceCrescent

Comrade.ConvolvedModel

Comrade.DFTAlg

Comrade.DFTAlg

Comrade.DesignMatrix

Comrade.Disk

Comrade.EHTArrayConfiguration

Comrade.EHTClosurePhaseDatum

Comrade.EHTLogClosureAmplitudeDatum

Comrade.EHTObservation

Comrade.EHTVisibilityAmplitudeDatum

Comrade.EHTVisibilityDatum

Comrade.ExtendedRing

Comrade.GainCache

Comrade.GainCache

Comrade.GainModel

Comrade.GainPrior

Comrade.GainPrior

Comrade.Gaussian

Comrade.GeometricModel

Comrade.MRing

Comrade.ModelImage

Comrade.MultiRadioLikelihood

Comrade.ParabolicSegment

Comrade.PolarizedModel

Comrade.Posterior

Comrade.RIMEModel

Comrade.RadioLikelihood

Comrade.RenormalizedModel

Comrade.Ring

Comrade.RotatedModel

Comrade.Scan

Comrade.ScanTable

Comrade.ShiftedModel

Comrade.StretchedModel

Comrade.TransformedPosterior

Base.:+

Comrade.Crescent

Comrade.added

Comrade.amplitude

Comrade.amplitude

Comrade.amplitude

Comrade.amplitudes

Comrade.arrayconfig

Comrade.baselines

Comrade.basemodel

Comrade.bispectra

Comrade.bispectrum

Comrade.closure_phase

Comrade.closure_phase

Comrade.closure_phases

Comrade.coherencymatrix

Comrade.components

Comrade.convolved

Comrade.corrupt

Comrade.create_cache

Comrade.extract_amp

Comrade.extract_cphase

Comrade.extract_lcamp

Comrade.extract_vis

Comrade.fileio_load

Comrade.fileio_save

Comrade.fouriermap

Comrade.getdata

Comrade.getuv

Comrade.getuvtimefreq

Comrade.load_ehtim

Comrade.logclosure_amplitude

Comrade.logclosure_amplitudes

Comrade.make_pullback

Comrade.mbreve

Comrade.modelimage

Comrade.posangle

Comrade.prior_sample

Comrade.rad2μas

Comrade.renormed

Comrade.rescaleuv!

Comrade.rotated

Comrade.scale_image

Comrade.scale_uv

Comrade.scantable

Comrade.shifted

Comrade.smoothed

Comrade.stations

Comrade.stretched

Comrade.transform_image

Comrade.transform_uv

Comrade.update_cache

Comrade.uviterator

Comrade.uvpositions

Comrade.visibilities

Comrade.visibility

Comrade.μas2rad

ComradeBase.evpa

ComradeBase.intensitymap

ComradeBase.intensitymap!

ComradeBase.m̆

HypercubeTransform.ascube

HypercubeTransform.asflat

ParameterHandling.flatten

StatsBase.sample

TransformVariables.inverse

TransformVariables.transform

Model Definitions

Calibration Models

Comrade.corrupt — Function

corrupt(vis::AbstractArray, cache::GainCache, gains::AbstractArray)

Corrupt the visibilities vis with the gains gains using a cache.

This returns an array of corrupted visibilties. This is called internally by the GainModel when producing the visibilties.

Comrade.DesignMatrix — Type

struct DesignMatrix{X, M<:AbstractArray{X, 2}, T, S} <: AbstractArray{X, 2}

Internal type that holds the gain design matrices for visibility corruption. See GainCache.

Comrade.GainCache — Type

struct GainCache{D1<:Comrade.DesignMatrix, D2<:Comrade.DesignMatrix, T, S}

Fields

m1

Gain design matrix for the first station

m2

Gain design matrix for the second station

times

Set of times for each gain

stations

Set of stations for each gain

Notes

Internal type. This should be created using the GainCache(st::ScanTable) method.

Comrade.GainCache — Method

GainCache(st::ScanTable)

Creates a cache for the application of gain corruptions to the model visibilities. This cache consists of the gain design matrices for each station and the set of times and stations for each gain.

Comrade.GainModel — Type

struct GainModel{C, G<:AbstractArray, M} <: Comrade.RIMEModel

A model that applies gain corruptions to a Comrade model. This obeys the usual Comrade interface and can be evaluated using visibilities.

Fields

cache

Cache for the application of gain. This can be constructed with GainCache.

gains

Array of the specific gains that are to be applied to the visibilities.

model

Base model that will be used to compute the uncorrupted visibilities.

Comrade.GainPrior — Type

GainPrior(dists, st::ScanTable)

Creates a distribution for the gain priors for scan table st. The dists should be a NamedTuple of Distributions, where each name corresponds to a telescope or station in the scan table. See scantable. The resulting type if a subtype of the Distributions.AbstractDistribution so the usual Distributions interface should work.

Example

For the 2017 observations of M87 a common GainPrior call is:

julia> gdist = GainPrior((AA = LogNormal(0.0, 0.1),
                   AP = LogNormal(0.0, 0.1),
                   JC = LogNormal(0.0, 0.1),
                   SM = LogNormal(0.0, 0.1),
                   AZ = LogNormal(0.0, 0.1),
                   LM = LogNormal(0.0, 1.0),
                   PV = LogNormal(0.0, 0.1)
                ), st)

julia> x = rand(gdist)
julia> logdensityof(gdist, x)

Comrade.GainPrior — Method

GainPrior(dists, st::ScanTable)

Creates a distribution for the gain priors for scan table st. The dists should be a NamedTuple of Distributions, where each name corresponds to a telescope or station in the scan table. See scantable. The resulting type if a subtype of the Distributions.AbstractDistribution so the usual Distributions interface should work.

Example

For the 2017 observations of M87 a common GainPrior call is:

julia> gdist = GainPrior((AA = LogNormal(0.0, 0.1),
                   AP = LogNormal(0.0, 0.1),
                   JC = LogNormal(0.0, 0.1),
                   SM = LogNormal(0.0, 0.1),
                   AZ = LogNormal(0.0, 0.1),
                   LM = LogNormal(0.0, 1.0),
                   PV = LogNormal(0.0, 0.1)
                ), st)

julia> x = rand(gdist)
julia> logdensityof(gdist, x)

Comrade.RIMEModel — Type

abstract type RIMEModel <: ComradeBase.AbstractModel

Abstract type that encompasses all RIME style corruptions.

Combinators

Base.:+ — Method

Base.:+(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(m1+m2, 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

Comrade.added — Function

added(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(added(m1,m2), 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

Comrade.convolved — Function

convolved(m1::AbstractModel, m2::AbstractModel)

Convolve two models to create a composite ConvolvedModel.

julia> m1 = Ring()
julia> m2 = Disk()
julia> convolved(m1, m2)

Comrade.components — Function

components(m::AbstractModel)

Returns the model components for a composite model. This will return a Tuple with all the models you have constructed.

Example

julia> m = Gaussian() + Disk()
julia> components(m)
(Gaussian{Float64}(), Disk{Float64}())

Comrade.smoothed — Function

smoothed(m::AbstractModel, σ::Number)

Smooths a model m with a Gaussian kernel with standard deviation σ.

Notes

This uses convolved to created the model, i.e.

julia> m1 = Disk()
julia> m2 = Gaussian()
julia> convolved(m1, m2) == smoothed(m1, 1.0)
true

Comrade.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ComradeBase.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

ComradeBase.intensity_point if model intensity is IsAnalytic

intensitymap! if model intensity is NotAnalytic

intensitymap if model intensity is NotAnalytic

flux

radialextent

visibilities (optional)

Comrade.AddModel — Type

struct AddModel{T1, T2} <: Comrade.CompositeModel{T1, T2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call added or Base.+ when constructing a model

Example

julia> m1 = Disk() + Gaussian()
julia> m2 = added(Disk(), Gaussian()) + Ring()

Comrade.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: Comrade.CompositeModel{M1, M2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call convolved. Also see smoothed(m, σ) for a simplified function that convolves a model m with a Gaussian with standard deviation σ.

GeometricModels

Comrade.scale_uv — Function

scale_image(model::AbstractModifier, u, u)

Returns a number on how to scale the image visibility at u v for an modified model

Comrade.scale_image — Function

scale_image(model::AbstractModifier, x, y)

Returns a number of how to to scale the image intensity at x y for an modified model

Comrade.transform_uv — Function

transform_uv(model::AbstractModifier, u, u)

Returns a transformed u and v according to the model modifier

Comrade.transform_image — Function

transform_image(model::AbstractModifier, x, y)

Returns a transformed x and y according to the model modifier

Comrade.GeometricModel — Type

abstract type GeometricModel <: ComradeBase.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain. As a result a user only needs to implement the following methods

visibility_point

intensity_point

radialextent

Note that if the geometric model isn't analytic then the usual methods listed in Comrade.AbstractModel for non-analytic models need to be implemented.

Comrade.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: Comrade.GeometricModel

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

router

Outer radius of the crescent

rinner

Inner radius of the crescent (i.e. inside this radius there is a hole)

shift

Displacment of the inner disk radius

slash

Strength of the linear slash. Note that s∈[0.0,1.0] to ensure positivity in the image.

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using Comrade.jl's model composition functionality.

Comrade.Crescent — Function

Creates a Kamruddin and Dexter crescent model. This works by composing two disk models together.

Arguments

router: The radius of the outer disk

rinner: The radius of the inner disk

shift: How much the inner disk radius is shifted (positive is to the right)

floor: The floor of the inner disk 0 means the inner intensity is zero and 1 means it is a large disk.

Comrade.Disk — Type

Disk{T}() where {T}

Tophat disk geometrical model, i.e. the intensity profile

\[ I(x,y) = \begin{cases} \pi^{-1} & x^2+y^2 < 1 \\ 0 & x^2+y^2 \geq 0 \end{cases}\]

i.e. a unit radius and unit flux disk.

By default if T isn't given, Disk defaults to Float64

Comrade.ExtendedRing — Type

struct ExtendedRing{F<:Number} <: Comrade.GeometricModel

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

The formula in the image domain is given by

I(r,θ) = βᵅrᵅ⁻²exp(-β/r)/2πΓ(α)

where α = shape and β = shape+1

Note

We mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

shape

shape of the radial distribution

Note that if T isn't specified at construction then it defaults to Float64.

Comrade.Gaussian — Type

struct Gaussian{T} <: Comrade.GeometricModel

Gaussian with unit standard deviation and flux.

By default if T isn't given, Gaussian defaults to Float64

Comrade.MRing — Type

struct MRing{T, V<:AbstractArray{T, 1}} <: Comrade.GeometricModel

m-ring geometric model. This is a infinitely thin unit flux delta ring whose angular structure is given by a Fourier expansion. That is,

I(r,θ) = (2π)⁻¹δ(r-1)∑ₙ(αₙcos(nθ) - βₙsin(nθ))

The N in the type defines the order of the Fourier expansion.

Fields

α

Real Fourier mode coefficients

β

Imaginary Fourier mode coefficients

Comrade.Ring — Type

struct Ring{T} <: Comrade.GeometricModel

A infinitely thin ring model, whose expression in the image domain is I(r,θ) = δ(r - 1)/2π i.e. a unit radius and flux delta ring.

By default if T isn't given, Gaussian defaults to Float64

Comrade.ParabolicSegment — Type

struct ParabolicSegment{T} <: Comrade.GeometricModel

A infinitely thin parabolic segment in the image domain. The segment is centered at zero, with roots ±1 and a yintercept of 1.

Note that if T isn't specified at construction then it defaults to Float64.

Model Image (non analytic FFT)

Comrade.create_cache — Function

create_cache(alg::AbstractFourierTransform, img::AbstractIntensityMap)

Creates a Fourier transform cache for a img using algorithm alg. For non-analytic visibility models this can significantly speed up computations.

Examples

julia> u,v = rand(100), rand(100)
julia> cache = create_cache(DFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(NFFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(FFTAlg(), IntensityMap(randn(50,50), 10.0, 10.0))

Comrade.update_cache — Function

update_cache(cache, img)

Update the Fourier transform cache. This will reuse an FFT/NFFT plan saving some computational time.

Note

This is an intenal method than an end user shouldn't have to usually call.

Comrade.modelimage — Function

modelimage(model::AbstractIntensityMap, image::AbstractIntensityMap, alg=FFTAlg(), executor=SequentialEx())

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg and executor which uses Folds.jl parallelism formalism.

Notes

For analytic models this is a no-op and returns the model. For non-analytic models this creates a ModelImage object which uses alg to compute the non-analytic Fourier transform.

modelimage(model, cache::AbstractCache, executor=SequentialEx())

Construct a ModelImage from the model and using a precompute Fourier transform cache. You can optionally specify the executor which will compute the internal image buffer using the executor.

Example

julia> m = ExtendedRing(10.0)
julia> cache = create_cache(DFTAlg(), IntensityMap(zeros(128, 128), 50.0, 50.0)) # used threads to make the image
julia> mimg = modelimage(m, cache, ThreadedEx())

Notes

For analytic models this is a no-op and returns the model.

modelimage(img::IntensityMap, alg=NFFTAlg())

Create a model image directly using an image, i.e. treating it as the model. You can optionally specify the Fourier transform algorithm using alg

Notes

For analytic models this is a no-op and returns the model.

modelimage(img::IntensityMap, cache::AbstractCache)

Create a model image directly using an image, i.e. treating it as the model. Additionally reuse a previously compute image cache. This can be used when directly modeling an image of a fixed size and number of pixels.

Notes

For analytic models this is a no-op and returns the model.

modelimage(m;
           fovx=2*radialextent(m),
           fovy=2*radialextent(m),
           nx=512,
           ny=512,
           pulse=ComradeBase.DeltaPulse(),
           alg=FFTAlg(),
           executor=SequentialEx()
            )

Construct a ModelImage where just the model m is specified.

If fovx or fovy aren't given modelimage will guess a reasonable field of view based on the radialextent function. nx and ny are the number of pixels in the x and y direction. The pulse is the pulse used for the image and alg

Notes

For analytic models this is a no-op and returns the model.

Comrade.uviterator — Function

uviterator(dx, dy, nnx, nny)

Construct the u,v iterators for the Fourier transform of the image with pixel sizes dx, dy and number of pixels nx, ny

If you are extending Fourier transform stuff please use these functions to ensure that the centroid is being properly computed.

Comrade.fouriermap — Function

fouriermap(m, fovx, fovy, nx, ny)

Create a Fourier or visibility map of a model m assuming a image with a field of view fovx/fovy and nx/ny pixels in the x/y direction respectively.

Comrade.ModelImage — Type

struct ModelImage{M, I, C} <: Comrade.AbstractModelImage{M}

Container for non-analytic model that contains a image cache which will hold the image, a Fourier transform cache, which usually an instance of a <: FourierCache.

Note

This is an internal implementation detail that shouldn't usually be called directly. Instead the user should use the exported function modelimage, for example

using Comrade
m = ExtendedRing(20.0, 5.0)

# This creates an version where the image is dynamically specified according to the
# radial extent of the image
mimg = modelimage(m) # you can also optionally pass the number of pixels nx and ny

# Or you can create an IntensityMap
img = intensitymap(m, 100.0, 100.0, 512, 512)
mimg = modelimage(m, img)

# Or precompute a cache
cache = create_cache(FFTAlg(), img)
mimg = modelimage(m, cache)

Comrade.DFTAlg — Type

DFTAlg

Uses a discrete fourier transform. This is not very efficient for larger images. In those cases NFFTAlg or FFTAlg are more reasonable. For small images this is a reasonable choice especially since it's easy to define derivatives.

Comrade.DFTAlg — Method

DFTAlg(obs::EHTObservation)

Create an algorithm object using the direct Fourier transform object from the observation obs. This will extract the uv positions from the observation to allow for a more efficient FT cache.

Modifiers

Comrade.basemodel — Function

basemodel(model::AbstractModel)

Returns the base model from a modified model. If there is no basemodel this just return the model itself.

Example

julia> basemodel(stretched(Disk(), 1.0, 2.0)) == Disk()
true

Comrade.renormed — Function

renormed(model, f)

Renormalizes the model m to have total flux f*flux(m). This can also be done directly by calling Base.:* i.e.,

julia> renormed(m, f) == f*M
true

Comrade.rotated — Function

rotated(model, ξ)

Rotates the model by an amount ξ in radians in the clockwise direction.

Comrade.posangle — Function

posangle(model)

Returns the rotation angle of the rotated model

Comrade.shifted — Function

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy in the x and y directions respectively.

Comrade.stretched — Function

stretched(model, α, β)

Stretches the model m according to the formula Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

Comrade.AbstractModifier — Type

abstract type AbstractModifier{M<:ComradeBase.AbstractModel} <: ComradeBase.AbstractModel

Abstract type for image modifiers. These are some model wrappers that can transform any model using simple Fourier transform properties. By default these modified models will have the same analytic properties as the base unmodified model, i.e.

julia> visanalytic(stretched(Disk(), 2.0, 2.0)) == visanalytic(Disk())
true

Additionally these are classic examples of non-primitive images i.e.,

julia> isprimitive(Comrade.AbstractModifier) == Comrade.NotAnalytic()

As a result of this the implementation of a model is slightly different

transform_uv

transform_image

scale_uv

scale_image

radialextent

This methods assume the modifiers are of the form

I(x,y) -> fᵢ(x,y)I(gᵢ(x,y)) V(u,v) -> fᵥ(u,v)V(gᵥ(u,v))

where g are the transformimage/uv functions and f are the scaleimage/uv function.

See those docstrings for guidance on implementation details.

Comrade.RenormalizedModel — Type

struct RenormalizedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Renormalizes the flux of the model to the new value scale*flux(model). We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

An end user should not call this directly but instead the renormed function or Base.:* instead.

Example

julia> renormed(Gaussian(), 2.0) == 2.0*Gaussian()
true

Comrade.RotatedModel — Type

struct RotatedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Type for the rotated model. This is more fine grained constrol of rotated model.

An end user should not call this directly but instead the rotated function instead.

Comrade.ShiftedModel — Type

struct ShiftedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Shifts the model by Δx units in the x-direction and Δy units in the y-direction.

An end user should not call this directly but instead the shifted function instead.

Comrade.StretchedModel — Type

struct StretchedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Stretched the model in the x and y directions, i.e. the new intensity is Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

An end user should not call this directly but instead the stretched function instead.

Polarized Models

ComradeBase.m̆ — Method

m̆(pimg::AbstractPolarizedModel, u, v)

Computes the fractional linear polarization in the visibility domain

m̆ = (Q + iU)/I

To create the symbol type m\breve in the REPL or use the mbreve function.

Comrade.mbreve — Function

mbreve(pimg, u, v)

Explicit m̆ function used for convenience.

ComradeBase.evpa — Method

evpa(pimg::AbstractPolarizedModel, u, v)

electric vector position angle or EVPA of the polarized model pimg at u and v

Comrade.coherencymatrix — Method

coherencymatrix(pimg, u, v)

Computes the coherency matrix of the polarized model pimg at u and v

Comrade.PolarizedModel — Type

struct PolarizedModel{TI, TQ, TU, TV} <: ComradeBase.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

Fields

I

Stokes I model

Q

Stokes Q Model

U

Stokes U Model

V

Stokes V Model

Model Evaluation

For more docstrings on how to evaluate models see ComradeBase API.

Comrade.amplitude — Function

amplitude(d::EHTVisibilityDatum)

Get the amplitude of a visibility datum

amplitude(d::EHTVisibilityAmplitudeDatum)

Get the amplitude of a amplitude datum

amplitude(model, args...)

Computes the visibility amplitude of model m at u,v positions u,v

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function.

Comrade.amplitudes — Function

amplitudes(m::AbstractModel, u::AbstractArray, v::AbstractArray)

Computes the amplitudes of the model m at the u,v positions u, v.

Comrade.bispectra — Function

bispectra(m,
          u1::AbstractArray, v1::AbstractArray,
          u2::AbstractArray, v2::AbstractArray,
          u3::AbstractArray, v3::AbstractArray,
        )

Computes the bispectra of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Comrade.bispectrum — Function

bispectrum(d1::T, d2::T, d3::T) where {T<:EHTVisibilityDatum}

Finds the bispectrum of three visibilities. We will assume these form closed triangles, i.e. the phase of the bispectrum is a closure phase.

bispectrum(model, u1, v1, u2, v2, u3, v3)

Computes the complex bispectrum of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function.

Comrade.closure_phase — Function

closure_phase(D1::EHTVisibilityDatum,
              D2::EHTVisibilityDatum,
              D3::EHTVisibilityDatum
              )

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

closure_phase(model. u1, v1, u2, v2, u3, v3)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function.

Comrade.closure_phases — Function

closure_phases(m,
               u1::AbstractArray, v1::AbstractArray,
               u2::AbstractArray, v2::AbstractArray,
               u3::AbstractArray, v3::AbstractArray,
               )

Computes the closure phases of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

closure_phases(m::AbstractModel, ac::ClosureConfig)

Computes the closure phases of the model m using the array configuration ac.

Notes

This is faster than the closure_phases(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

closure_phases(vis::AbstractArray, ac::ArrayConfiguration)

Compute the closure phases for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

Comrade.logclosure_amplitude — Function

logclosure_amplitude(model, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v4 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function.

Comrade.logclosure_amplitudes — Function

logclosure_amplitudes(m::AbstractModel,
                      u1::AbstractArray, v1::AbstractArray,
                      u2::AbstractArray, v2::AbstractArray,
                      u3::AbstractArray, v3::AbstractArray,
                      u4::AbstractArray, v4::AbstractArray,
                     )

Computes the log closure amplitudes of the model m at the quadrangles (u1,v1), (u2,v2), (u3,v3), (u4, v4).

logclosure_amplitudes(m::AbstractModel, ac::ClosureConfig)

Computes the log closure amplitudes of the model m using the array configuration ac.

Notes

This is faster than the logclosure_amplitudes(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

logclosure_amplitudes(vis::AbstractArray, ac::ArrayConfiguration)

Compute the log-closure amplitudes for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

Comrade.visibilities — Function

visibilities(m, ac::ArrayConfiguration)

Computes the visibilities of the model m using the array configuration ac. If you want to compute a single visibility you should call visibility.

visibilities(m, u::AbstractArray, v::AbstractArray, args...)

Computes the visibilities of the model m at u v and args.... If you want to compute a single visibility you should call visibility.

Comrade.visibility — Function

visibility(d::EHTVisibilityDatum)

Return the complex visibility of the visibility datum

visibility(mimg, u, v, args...)

Computes the complex visibility of model m at u,v positions u,v and other args... (e.g., frequency, time, etc.)

Notes

If you want to compute the visibilities at a large number of positions consider using the visibilities.

visibility(mimg, uv::ArrayBaselineDatum)

Computes the complex visibility of a model m at the uv array baseline datum.

Notes

If you want to compute the visibilities at a large number of positions consider using the visibilities.

visibility(pimg::PolarizedModel, u, v)

Computes the visibility in the stokes basis of the polarized model

ComradeBase.intensitymap — Function

intensitymap(model::AbstractModel, args...)

Computes the intensity map of model. For the inplace version see intensitymap!

ComradeBase.intensitymap! — Function

intensitymap!(buffer::AbstractMatrix, model::AbstractModel, args...)

Computes the intensity map of model by modifying the buffer

Data Types

Comrade.amplitude — Method

amplitude(d::EHTVisibilityDatum)

Get the amplitude of a visibility datum

Comrade.amplitude — Method

amplitude(d::EHTVisibilityAmplitudeDatum)

Get the amplitude of a amplitude datum

amplitude(model, args...)

Computes the visibility amplitude of model m at u,v positions u,v

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function.

Comrade.baselines — Function

baselines(CP::EHTClosurePhaseDatum)

Returns the baselines used for a single closure phase datum

baselines(CP::EHTLogClosureAmplitudeDatum)

Returns the baselines used for a single closure phase datum

baselines(scan::Scan)

Return the baselines for each datum in a scan

Comrade.arrayconfig — Function

arrayconfig(vis)

Extract the array configuration from a EHT observation.

Comrade.closure_phase — Method

closure_phase(D1::EHTVisibilityDatum,
              D2::EHTVisibilityDatum,
              D3::EHTVisibilityDatum
              )

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

Comrade.getdata — Function

getdata(obs::EHTObservation, s::Symbol)

Pass-through function that gets the array of s from the EHTObservation. For example say you want the times of all measurement then

getdata(obs, :time)

Comrade.getuv — Function

getuv

Get the u, v positions of the array.

Comrade.getuvtimefreq — Function

getuvtimefreq(ac)

Get the u, v, time, freq of the array as a tuple.

Comrade.rescaleuv! — Function

rescaleuv!(data::EHTObservation)

rescale the u-v lengths according to scale. This can be useful when you want the spatial scales to be in 1/μas instead of 1/rad

Comrade.scantable — Function

scantable(obs::EHTObservation)

Reorganizes the observation into a table of scans, where scan are defined by unique timestamps. To access the data you can use scalar indexing

Example

st = scantable(obs)
# Grab the first scan
scan1 = st[1]

# Acess the detections in the scan
scan1[1]

# grab e.g. the baselines
scan1[:baseline]

Comrade.stations — Function

stations(d::EHTObservation)

Get all the stations in a observation. The result is a vector of symbols.

stations(g::CalTable)

Return the stations in the calibration table

Comrade.uvpositions — Function

uvpositions(datum::AbstractVisibilityDatum)

Get the uvp positions of an inferometric datum.

Comrade.ClosureConfig — Type

struct ClosureConfig{A, D} <: Comrade.ArrayConfiguration

Array config file for closure quantities. This stores the design matrix designmat that transforms from visibilties to closure products.

Fields

ac

Array configuration for visibilities

designmat

Closure design matrix

Comrade.ArrayBaselineDatum — Type

struct ArrayBaselineDatum{T}

A single datum of an ArrayConfiguration

Comrade.EHTObservation — Type

struct EHTObservation{F, T<:Comrade.AbstractInterferometryDatum{F}, S<:(StructArrays.StructArray{T<:Comrade.AbstractInterferometryDatum{F}}), A, N} <: Comrade.Observation{F}

The main data product type in Comrade this stores the data which can be a StructArray of any AbstractInterferometryDatum type.

Fields

data

StructArray of data productts

config

Array config holds ancillary information about array

mjd

modified julia date of the observation

ra

RA of the observation in J2000 (deg)

dec

DEC of the observation in J2000 (deg)

bandwidth

bandwidth of the observation (Hz)

frequency

frequency of the observation (Hz)

source

Common source name

timetype

Time zone used.

Comrade.EHTArrayConfiguration — Type

struct EHTArrayConfiguration{F, T<:AbstractArray} <: Comrade.ArrayConfiguration

Stores all the non-visibility data products for an EHT array. This is useful when evaluating model visibilities.

Fields

frequency

Observing frequency (Hz)

bandwidth

Observing bandwith (Hz)

data

A struct array of ArrayBaselineDatum holding time, freq, u, v, baselines.

Comrade.EHTClosurePhaseDatum — Type

struct EHTClosurePhaseDatum{T<:Number} <: Comrade.ClosureProducts{T<:Number}

A Datum for a single closure phase.

Fields

phase

closure phase (rad)

error

error of the closure phase assuming the high-snr limit

u1

u (λ) of first station

v1

v (λ) of first station

u2

u (λ) of second station

v2

v (λ) of second station

u3

u (λ) of third station

v3

v (λ) of third station

time

Measured time of closure phase in hours

triangle

station baselines used

Comrade.EHTLogClosureAmplitudeDatum — Type

struct EHTLogClosureAmplitudeDatum{T<:Number} <: Comrade.ClosureProducts{T<:Number}

A Datum for a single log closure amplitude.

amp

log-closure amplitude

error

log-closure amplitude error in the high-snr limit

u1

u (λ) of first station

v1

v (λ) of first station

u2

u (λ) of second station

v2

v (λ) of second station

u3

u (λ) of third station

v3

v (λ) of third station

u4

u (λ) of fourth station

v4

v (λ) of fourth station

time

Observation time of the quadrangle

quadrangle

station codes for the quadrangle

Comrade.EHTVisibilityDatum — Type

struct EHTVisibilityDatum{T<:Number} <: Comrade.AbstractVisibilityDatum{T<:Number}

A struct holding the information for a single measured visibility.

visr

real component of the visibility (Jy)

visi

imaginary component of the visibility (Jy)

error

error of the visibility (Jy)

u

x-direction baseline length in λ

v

y-direction baseline length in λ

time

Time of the observation in hours

baseline

station baseline codes

Comrade.EHTVisibilityAmplitudeDatum — Type

struct EHTVisibilityAmplitudeDatum{T<:Number} <: Comrade.AbstractVisibilityDatum{T<:Number}

A struct holding the information for a single measured visibility amplitude.

FIELDS

amp

amplitude (Jy)

error

error of the visibility amplitude (Jy)

u

x-direction baseline length in λ

v

y-direction baseline length in λ

time

Time of the observation in hours

baseline

station baseline codes

Comrade.Scan — Type

struct Scan{T, I, S}

Composite type that holds information for a single scan of the telescope.

Fields

time

Scan time

index

Scan indices which are (scan index, data start index, data end index)

scan

Scan data usually a StructArray of a <:AbstractVisibilityDatum

Comrade.ScanTable — Type

struct ScanTable{O<:Union{Comrade.ArrayConfiguration, Comrade.Observation}, T, S}

Wraps EHTObservation in a table that separates the observation into scans. This implements the table interface. You can access scans by directly indexing into the table. This will create a view into the table not copying the data.

Example

julia> st = scantable(obs)
julia> st[begin] # grab first scan
julia> st[end]   # grab last scan

eht-imaging interface

Comrade.extract_amp — Function

extract_amp(obs)

Extracts the visibility amplitudes from an ehtim observation object.

Any valid keyword arguments to add_amp in ehtim can be passed through extract_amp.

Returns an EHTObservation with visibility amplitude data

Comrade.extract_cphase — Function

extract_cphase(obs)

Extracts the closure phases from an ehtim observation object

Any valid keyword arguments to add_cphase in ehtim can be passed through extract_cphase.

Returns an EHTObservation with closure phases datums

Special Keyword arguments:

count: How the closures are formed, the available options are "min-correct", "min", "max"

cut_trivial: Cut the trivial triangles from the closures

uvmin: The flag to decide what are trivial triangles. Any baseline with ||(u,v)|| < uvmin are removed.

kwargs...: Other arguments are forwarded to eht-imaging.

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_lcamp — Function

extract_lcamp(obs)

Extracts the log-closure amp. from an ehtim observation object

Any valid keyword arguments to add_logcamp in ehtim can be passed through extract_lcamp.

Special Keyword arguments:

count: How the closures are formed, the available options are "min-correct", "min", "max"

kwargs...: Other arguments are forwarded to eht-imaging.

Returns an EHTObservation with log-closure amp. datums

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_vis — Function

extract_vis(obs)

Extracts the complex visibilities from an ehtim observation object

This grabs the raw data object from the obs object. Any keyword arguments are ignored.

Returns an EHTObservation with complex visibility data

Comrade.load_ehtim — Function

load_ehtim()

Loads the eht-imaging library and stores it in the exported ehtim variable.

Notes

This will fail if ehtim isn't installed in the python installation that PyCall references.

Bayesian Tools

Distributions

Comrade.AmpNormal — Type

AmpNormal

Visibility amplitude likelihood distribution. Typically μ, τ=1/σ are vectors to allow for more efficient derivatives.

Comrade.ComplexNormal — Type

ComplexNormal

Uncorrelated complex Normal Measure. This is the default likelihood used for visibilties.

Comrade.CPVonMises — Type

CPVonMises

The von Mises distribution used for closure phases. Typically μ and κ=1/σ² are vectors to allow for simpler derivative rules.

Posterior Constructions

HypercubeTransform.ascube — Function

ascube(post::Posterior)

Construct a flattened version of the posterior where the parameters are transformed to live in (0, 1), i.e. the unit hypercube.

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = ascube(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical NestedSampling methods, i.e. ComradeNested. For the transformation to unconstrained space see asflat

HypercubeTransform.asflat — Function

asflat(post::Posterior)

Construct a flattened version of the posterior where the parameters are transformed to live in (-∞, ∞).

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = ascube(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical MCMC methods, i.e. ComradeAHMC. For the transformation to the unit hypercube see ascube

ParameterHandling.flatten — Function

flatten(post::Posterior)

Construct a flattened version of the posterior but do not transform to any space, i.e. use the support specified by the prior.

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = flatten(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical MCMC methods, i.e. ComradeAHMC. For the transformation to the unit hypercube see ascube

TransformVariables.inverse — Function

inverse(posterior::TransformedPosterior, x)

Transforms the value y from parameter space to the transformed space (e.g. unit hypercube if using ascube).

For the inverse transform see transform

Comrade.prior_sample — Function

prior_sample(post::Posterior, args...)

Samples the prior distribution from the posterior. The args... are forwarded to the Base.rand method.

prior_sample(post::Posterior)

Returns a single sample from the prior distribution.

StatsBase.sample — Method

sample(post::Posterior, sampler::S, args...; init_params=nothing, kwargs...)

Sample a posterior post using the sampler. You can optionally pass the starting location of the sampler using init_params, otherwise a random draw from the prior will be used.

TransformVariables.transform — Function

transform(posterior::TransformedPosterior, x)

Transforms the value x from the transformed space (e.g. unit hypercube if using ascube) to parameter space which is usually encoded as a NamedTuple.

For the inverse transform see inverse

Comrade.MultiRadioLikelihood — Type

MultiRadioLikelihood(lklhd1, lklhd2, ...)

Combines multiple likelihoods into one object that is useful for fitting multiple days/frequencies.

julia> lklhd1 = RadioLikelihood(dcphase1, dlcamp1)
julia> lklhd2 = RadioLikelihood(dcphase2, dlcamp2)
julia> MultiRadioLikelihood(lklhd1, lklhd2)

Comrade.Posterior — Type

Posterior(lklhd, prior, model)

Creates a Posterior density that follows obeys DensityInferface. The lklhd object is expected to be a MeasureBase.Likelihood object. For instance, these can be created using RadioLikelihood. prior is expected to be a NamedTuple of distributions that reflect the priors on the parameters you are considering. model is a function that takes in a NamedTuple of parameters and returns a Comrade <:AbstractModel.

Notes

Since this function obeys DensityInferface you can evaluate it with

julia> ℓ = logdensityof(post)
julia> ℓ(x)

or using the 2-argument version directly

julia> logdensityof(post, x)

where post::Posterior.

To generate random draws from the prior see the prior_sample function.

Comrade.TransformedPosterior — Type

struct TransformedPosterior{P<:Posterior, T}

A transformed version of a Posterior object. This is an internal type that an end user shouldn't have to directly construct. To construct a transformed posterior see the asflat, ascube, and flatten docstrings.

Comrade.RadioLikelihood — Type

RadioLikelihood(data1, data2, ...)

Forms a radio likelihood from a set of data products. These data products must share the same array data/configuration. If you want to form a likelihood from multiple arrays such as when fitting different wavelengths or days, you can combine them using MultiRadioLikelihood

Example

julia> RadioLikelihood(dcphase1, dlcamp1)

Misc

Comrade.μas2rad — Function

μas2rad(x)

Converts a number from micro-arcseconds (μas) to rad

Comrade.rad2μas — Function

rad2μas(x)

Converts a number from radians to micro-arcseconds (μas)

Comrade.fileio_load — Function

load(fitsfile::String, IntensityMap)

This loads in a fits file that is more robust to the various imaging algorithms in the EHT, i.e. is works with clean, smili, eht-imaging. The function returns an tuple with an intensitymap and a second named tuple with ancillary information about the image, like the source name, location, mjd, and radio frequency.

Comrade.fileio_save — Function

save(file::String, img::IntensityMap, obs)

Saves an image to a fits file. You can optionally pass an EHTObservation so that ancillary information will be added.

Comrade.make_pullback — Function

make_pullback(ℓ, autodiff::AD.AbstractBackend)

Create the pullback function using the autodiff backend autodiff.

Note

This is an internal function and not part of the public API.

make_pullback(ℓ, grad::Function)

Create the pullback function using the function grad which should return the gradient of ℓ.

Note

This is an internal function and shouldn't be typically used by an end-user.

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020