ComradeBase API

Contents

ComradeBase API

Contents

Index

Index

ComradeBase.BSplinePulse

ComradeBase.CoherencyMatrix

ComradeBase.DeltaPulse

ComradeBase.DensityAnalytic

ComradeBase.IntensityMap

ComradeBase.IsAnalytic

ComradeBase.IsPrimitive

ComradeBase.NotAnalytic

ComradeBase.NotPrimitive

ComradeBase.PrimitiveTrait

ComradeBase.Pulse

ComradeBase.SingleStokes

ComradeBase.SqExpPulse

ComradeBase.StokesVector

ComradeBase.evpa

ComradeBase.imanalytic

ComradeBase.intensity_point

ComradeBase.intensitymap

ComradeBase.intensitymap!

ComradeBase.isprimitive

ComradeBase.m̆

ComradeBase.visanalytic

ComradeBase.visibility_point

ComradeBase.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into Comrade more easily.

ComradeBase.IsPrimitive — Type

struct IsPrimitive

Trait for primitive model

ComradeBase.NotPrimitive — Type

struct NotPrimitive

Trait for not-primitive model

ComradeBase.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive Comrade model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of Comrade they need to specify if it is primitive

struct MyPrimitiveModel end
ComradeBase.isprimitive(::Type{MyModel}) = ComradeBase.IsPrimitive()

ComradeBase.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ComradeBase.IsAnalytic — Type

struct IsAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ComradeBase.NotAnalytic — Type

struct NotAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ComradeBase.visanalytic — Function

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ComradeBase.imanalytic — Function

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ComradeBase.visibility_point — Function

Function that computes the pointwise visibility if the model has the trait in the fourier domain IsAnalytic(). Otherwise it will use the FFTW fallback.

ComradeBase.intensity_point — Function

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ComradeBase.IntensityMap — Type

struct IntensityMap{T, S<:(AbstractMatrix), F, K<:ComradeBase.Pulse} <: ComradeBase.AbstractIntensityMap{T, S<:(AbstractMatrix)}

Image array type. This is an Matrix with a number of internal fields to describe the field of view, pixel size, and the pulse function that makes the image a continuous quantity.

To use it you just specify the array and the field of view/pulse $julia img = IntensityMap(zeros(512,512), 100.0, 100.0)$`

ComradeBase.intensitymap! — Function

Computes the intensity map of model by modifying the input IntensityMap object

ComradeBase.intensitymap — Function

Computes the intensity map of model. This version requires additional information to construct the grid.

Example

m = Gaussian()
# field of view
fovx, fovy = 5.0
fovy = 5.0
# number of pixels
nx, ny = 128

img = intensitymap(m, fovx, fovy, nx, ny; pulse=DeltaPulse())

ComradeBase.Pulse — Type

Pulse Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

ComradeBase.DeltaPulse — Type

struct DeltaPulse{T} <: ComradeBase.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

ComradeBase.SqExpPulse — Type

struct SqExpPulse{T} <: ComradeBase.Pulse

Normalized square exponential kernel, i.e. a Gaussian. Note the smoothness is modfied with ϵ which is the inverse variance in units of 1/pixels².

ComradeBase.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

ComradeBase.StokesVector — Type

struct StokesVector{T} <: StaticArrays.FieldVector{4, T}

Static vector that holds the stokes parameters of a polarized complex visibility

To convert between a StokesVector and CoherencyMatrix use the convert function

convert(::CoherencyMatrix, StokesVector(1.0, 0.1, 0.1, 0.4))

ComradeBase.CoherencyMatrix — Type

struct CoherencyMatrix{T} <: StaticArrays.FieldMatrix{2, 2, T}

Static matrix that holds construct the coherency matrix of a polarized complex visibility

To convert between a StokesVector and CoherencyMatrix use the convert function

convert(::StokesVector, CoherencyMatrix(1.0, 0.1, 0.1, 0.4))

ComradeBase.evpa — Function

evpa(pimg, u, v)

electric vector position angle or EVPA of the polarized model

evpa(m)

Compute the evpa of a stokes vector or cohereny matrix.

ComradeBase.m̆ — Function

m̆(pimg, u, v)

Computes the fractional linear polarization in the visibility domain

\[ \breve{m} = \frac{Q + iU}{I}\]

m̆(m)

Compute the fractional linear polarization of a stokes vector or coherency matrix

ComradeBase.SingleStokes — Type

struct SingleStokes{M, S} <: ComradeBase.AbstractModel

Helper function that converts a model from something that compute polarized images to just a single stokes parameter. This is useful if you just want to fit a single stokes parameter.

Missing docstring.

Missing docstring for ComradeBase.ThreadedModel. Check Documenter's build log for details.