ComradeNested

ComradeNested interfaces Comrade to the excellent NestedSamplers.jl package.

We follow NestedSamplers interface closely. The difference is that instead of creating a NestedModel, we pass a Comrade.Posterior object as our model. Internally, Comrade defines the prior transform and extracts the log-likelihood function.

For more information about NestedSamplers.jl please see its docs.

Example

using Comrade
using ComradeNested

# Some stuff to create a posterior object
post # of type Comrade.Posterior

# Create sampler using 1000 live points
smplr = Nested(dimension(post), 1000)

samples = sample(post, smplr; d_logz=1.0)

# Optionally resample the chain to create an equal weighted output
using StatsBase
equal_weight_chain = ComradeNested.equalresample(samples, 10_000)

API

StatsBase.sample — Method

AbstractMCMC.sample(post::Comrade.Posterior, smplr::Nested, args...; kwargs...)

Sample the posterior post using NestedSamplers.jl Nested sampler. The args/kwargs are forwarded to NestedSampler for more information see its docs

This returns a tuple where the first element are the weighted samples from NestedSamplers in a TypedTable. The second element includes additional information about the samples, like the log-likelihood, evidence, evidence error, and the sample weights.

To create equally weighted samples the user can use ```julia using StatsBase chain, stats = sample(post, NestedSampler(dimension(post), 1000)) equalweightedchain = sample(chain, Weights(stats.weights), 10_000)