API

Comrade.Comrade — Module

Comrade

Radio Observation Sampling Exploration

Comrade.AbstractCache — Type

abstract type AbstractCache

This defines an abstract cache that can be used to hold or precompute some computations.

Comrade.AbstractModifier — Type

abstract type AbstractModifier{M<:ComradeBase.AbstractModel} <: ComradeBase.AbstractModel

Abstract type for image modifiers. These are some model wrappers that can transform any model using simple Fourier transform properties. To see the implemented modifier

Comrade.AddModel — Type

struct AddModel{T1, T2} <: Comrade.CompositeModel{T1, T2}

Adds two models together to create composite models. Note that I may change this in the future so make it easier on the compiler, i.e. make the composite model a fancy model vector and heap allocate stuff. This should help when combining multiple models together.

Comrade.ArrayConfiguration — Type

abstract type ArrayConfiguration

This defined the abstract type for an array configuration. Namely, baseline times, SEFD's, bandwidth, observation frequencies, etc.

Comrade.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ComradeBase.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

ComradeBase.intensity_point if model intensity is IsAnalytic

intensitymap! if model intensity is NotAnalytic

flux

Comrade.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: Comrade.GeometricModel

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using Comrade.jl's model composition functionality.

Comrade.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: Comrade.CompositeModel{M1, M2}

Smooths a model with a Gaussian kernel with standard deviation of σ.

Notes

m = smoothed(Disk(), 5.0)
intensity(m, 2.0, 2.0; fov=10.0, npix=128)

will compute the intensity of the smoothed disk with using an interpolation with 128 pixel nodes and a total fov of 10 in x and y direction.

This needs to be improved

Comrade.EHTClosurePhaseDatum — Type

struct EHTClosurePhaseDatum{T<:Number} <: Comrade.ClosureProducts{T<:Number}

A Datum for a single closure phase. Note in the future this may get replaced with the fully covariant formalism from Blackburn et al. (2020).

Comrade.ExtendedRing — Type

struct ExtendedRing{F} <: Comrade.GeometricModel

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

Note

@e mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

radius

radius of peak emission

shape

shape of the radial distribution

Comrade.FFT — Type

struct FFT <: Comrade.FourierTransform

Fourier transform type that specifies we will use the FFTW package to compute the Fourier transform.

Fields

padfac

The amount to pad the image by. Note we actually round up to the nearest factor of 2, but this will be improved in the future to use small primes

Comrade.FFTCache — Type

struct FFTCache{P, I} <: Comrade.AbstractCache

The cache used when the FFT algorithm is used to compute visibilties. This is an internal type and is not part of the public API

Fields

plan

FFTW Plan

sitp

FFT interpolator function

Comrade.FourierTransform — Type

abstract type FourierTransform

Abstract type that specified which fourier transform to use

Comrade.Gaussian — Type

struct Gaussian{T} <: Comrade.GeometricModel

Gaussian geometrical model. This is a Gaussian with unit flux and standard deviation.

Notes

To change the Gaussian flux, and shape please use the modifier functions

Comrade.GeometricModel — Type

abstract type GeometricModel <: ComradeBase.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain.

Comrade.MRing — Type

struct MRing{T, N} <: Comrade.GeometricModel

m-ring geometric model. This corresponds to a delta ring with a fourier expansion in θ. The m in m-ring refers to the order of the Fourier expansion. The radius is unity.

Comrade.ModelImage — Type

struct ModelImage{M, I, C} <: Comrade.AbstractModelImage{M}

Container for non-analytic model that contains a image cache which will hold the image, a Fourier transform cache C, which usually an instance of a <: FourierCache which usually holds a interpolator that allows you to compute visibilities.

Notes

This is an internal implementation detail that shouldn't usually be called directly. Instead the user should use the exported function modelimage, for example

using Comrade
m = ExtendedRing(20.0, 5.0)

# This creates an version where the image is dynamically specified according to the
# radial extent of the image
mimg = modelimage(m) # you can also optionally pass the number of pixels nx and ny

# Or you can create an IntensityMap
img = intensitymap(m, 100.0, 100.0, 512, 512)
mimg = modelimage(m, img)

Comrade.NoCache — Type

struct NoCache <: Comrade.AbstractCache

No cache is used. This is typically used when the model is analytic in the Fourier domain.

Comrade.PolarizedModel — Type

struct PolarizedModel{TI, TQ, TU, TV} <: ComradeBase.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

Comrade.RImage — Type

$(TYPEDEF)

An image model given by a set of coefficients and a kernel response or basis function. This corresponds to a continous image defined by a finite set of points. The defined intensity is given by

\[ I(x,y) = \sum_{ij} c_{ij}κ(x-x_i)κ(y-y_i).\]

An important thing to note is that the $c_{ij}$ do not represent pixel intensities, i.e. the κ doesn't have to be an interpolating kernel.

Example

samples = rand(10,10)
model = RImage(samples, BSplineKernel{3})

Notes

This is defined in terms of pixel response, so the image size is 1μas. To resize the image use the scale function like with other models.

Fields
FIELDS

Comrade.RenormalizedModel — Type

struct RenormalizedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Renormalizes the flux of the model to the new value flux. We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

Comrade.RotatedModel — Type

struct RotatedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Type for the rotated model. This is more fine grained constrol of rotated model. In most use cases the end-user should be using the rotate method e.g.

rotate(model, ξ)

Comrade.ShiftedModel — Type

struct ShiftedModel{T, M<:ComradeBase.AbstractModel} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Shifts the model by Δx units in the x-direction and Δy units in the y-direction.

Comrade.StretchedModel — Type

struct StretchedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Stretched the model in the x and y directions, i.e. the new intensity is

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

Base.size — Method

size(model)

return the size of the coefficient matrix for model.

Comrade.Crescent — Method

Creates a Kamruddin and Dexter crescent model. This works by composing two disk models together.

Arguments

router: The radius of the outer disk

rinner: The radius of the inner disk

shift: How much the inner disk radius is shifted (positive is to the right)

floor: The floor of the inner disk 0 means the inner intensity is zero and 1 means it is a large disk.

Comrade._visibilities — Method

_visibilities(m, u, v)

Computes the visibilities of the model m at the u,v positions u, v.

Note this is done lazily so the visibility is only computed when accessed.

Comrade.amplitude — Method

amplitude(model, u, v)

Computes the visibility amplitude of model m at u,v positions u,v

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.amplitudes — Method

amplitudes(m, u, v)

Computes the amplitudes of the model m at the u,v positions u, v.

Note this is done lazily so the visibility is only computed when accessed.

Comrade.basemodel — Method

basemodel(model)

Returns the base model from a modified model.

Comrade.bispectra — Method

bispectra(m, u1, v1, u2, v2, u3, v3)

Computes the bispectra of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the bispectra is only computed when accessed.

Comrade.bispectrum — Method

bispectrum(model, u1, v1, u2, v2, u3, v3)

Computes the complex bispectrum of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.closure_phase — Method

closure_phase(model, u1, v1, u2, v2, u3, v3)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.closure_phase — Method

closure_phase(D1, D2, D3)

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

Comrade.closure_phases — Method

closure_phases(m, u1, v1, u2, v2, u3, v3)

Computes the closure phases of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the closure_phases is only computed when accessed.

Comrade.coherencymatrix — Method

coherencymatrix(pimg, u, v)

Computes the coherency matrix of the polarized model.

Comrade.components — Method

components(m)

Returns the components for a composite model. This will return a Tuple with all the models you have constructed.

Comrade.create_cache — Method

create_cache(alg, img)

Creates the model cache given for the algorithm alg using the model and a image cache image

Comrade.extract_cphase — Method

extract_cphase(obs)

Extracts the closure phases from an ehtim observation object

Returns an EHTObservation with closure phases datums

Comrade.extract_lcamp — Method

extract_lcamp(obs)

Extracts the log-closure amp. from an ehtim observation object

Returns an EHTObservation with closure amp. datums

Comrade.extract_vis — Method

extract_vis(obs)

Extracts the complex visibilities from an ehtim observation object

Returns an EHTObservation with complex visibility data

Comrade.load_ehtim — Method

load_ehtim()

Loads the eht-imaging library and stores it in the ehtim variable.

Notes

This will fail if ehtim isn't installed in the python installation that PyCall references.

Comrade.load_tpy — Method

load_tpy(file)

Load a ThemisPy style ascii EHT observation file.

Comrade.logclosure_amplitude — Method

logclosure_amplitude(model, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v3 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.logclosure_amplitudes — Method

logclosure_amplitudes(m, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log closure amplitudes of the model m at the quadrangles (u1,v1), (u2,v2), (u3,v3), (u4, v4).

Note this is done lazily so the log closure amplitude is only computed when accessed.

Comrade.modelimage — Method

modelimage(m)

Construct a ModelImage where just the model m is specified

Notes

If m IsAnalytic() is the visibility domain this is a no-op and just returns the model itself. Otherwise modelimage will guess a reasonable field of view based on the radialextent function. One can optionally pass the number of pixels nx and ny in each direction.

Comrade.modelimage — Method

modelimage(model, image; alg)
modelimage(m; fovx, fovy, nx, ny, alg)

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg

Notes

For analytic models this is a no-op and just return the model. For non-analytic models this wraps the model in a object with an image and precomputes the fourier transform using alg.

Comrade.rad2μas — Method

rad2μas(x)

Converts a number from radians to μas

Comrade.renormed — Method

renormed(model, f)

Renormalizes the model m to have total flux flux.

Comrade.rotated — Method

rotated(model, ξ)

Rotates the model by an amount ξ in radians.

Comrade.shifted — Method

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy.

Comrade.stretched — Method

stretched(model, α, β)

Stretches the model m according to the formula

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

Comrade.visibility — Method

visibility(pimg, u, v)

Computes the visibility in the stokes basis of the polarized model

Comrade.visibility — Method

visibility(mimg, u, v)

Computes the complex visibility of model m at u,v positions u,v

If you want to compute the visibilities at a large number of positions consider using the visibilities function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.μas2rad — Method

μas2rad(x)

Converts a number from μas to rad

ComradeBase.evpa — Method

evpa(pimg, u, v)

electric vector position angle or EVPA of the polarized model

ComradeBase.intensitymap! — Method

intensitymap!(pimg, pmodel)

Finds the polarized intensity map of the polarized model pmodel.

ComradeBase.m̆ — Method

m̆(pimg, u, v)

Computes the fractional linear polarization in the visibility domain

\[ \breve{m} = \frac{Q + iU}{I}\]