Comrade API

Contents

Comrade API

Contents

Index

Index

Comrade.AbstractCache

Comrade.AbstractInterferometryDatum

Comrade.AbstractModifier

Comrade.AddModel

Comrade.ArrayBaselineDatum

Comrade.ArrayConfiguration

Comrade.ClosureConfig

Comrade.CompositeModel

Comrade.ConcordanceCrescent

Comrade.ConvolvedModel

Comrade.DFTAlg

Comrade.EHTArrayConfiguration

Comrade.EHTClosurePhaseDatum

Comrade.EHTLogClosureAmplitudeDatum

Comrade.EHTObservation

Comrade.EHTVisibilityAmplitudeDatum

Comrade.EHTVisibilityDatum

Comrade.ExtendedRing

Comrade.FFTAlg

Comrade.FFTCache

Comrade.GainModel

Comrade.Gaussian

Comrade.GeometricModel

Comrade.MRing

Comrade.ModelImage

Comrade.MultiRadioLikelihood

Comrade.NFFTAlg

Comrade.NonAnalyticTest

Comrade.ParabolicSegment

Comrade.ParabolicSegment

Comrade.PolarizedModel

Comrade.Posterior

Comrade.RadioLikelihood

Comrade.RenormalizedModel

Comrade.Ring

Comrade.RotatedModel

Comrade.ShiftedModel

Comrade.StretchedModel

Comrade.TransformedPosterior

Comrade.Crescent

Comrade.amplitude

Comrade.amplitude

Comrade.amplitudes

Comrade.arrayconfig

Comrade.basemodel

Comrade.bispectra

Comrade.bispectrum

Comrade.closure_phase

Comrade.closure_phase

Comrade.closure_phases

Comrade.coherencymatrix

Comrade.components

Comrade.convolved

Comrade.create_cache

Comrade.create_cache

Comrade.extract_amp

Comrade.extract_cphase

Comrade.extract_lcamp

Comrade.extract_vis

Comrade.getuv

Comrade.load_ehtim

Comrade.load_tpy

Comrade.logclosure_amplitude

Comrade.logclosure_amplitudes

Comrade.modelimage

Comrade.modelimage

Comrade.prior_sample

Comrade.rad2μas

Comrade.renormed

Comrade.rescaleuv!

Comrade.rotated

Comrade.scantable

Comrade.shifted

Comrade.smoothed

Comrade.stations

Comrade.stretched

Comrade.update_cache

Comrade.uviterator

Comrade.uvpositions

Comrade.uvtimefreq

Comrade.visibility

Comrade.visibility

Comrade.visibility

Comrade.μas2rad

ComradeBase.evpa

ComradeBase.intensitymap!

ComradeBase.m̆

HypercubeTransform.ascube

HypercubeTransform.asflat

ParameterHandling.flatten

RecipesBase.apply_recipe

RecipesBase.apply_recipe

TransformVariables.inverse

TransformVariables.transform

Comrade.Crescent — Method

Creates a Kamruddin and Dexter crescent model. This works by composing two disk models together.

Arguments

router: The radius of the outer disk

rinner: The radius of the inner disk

shift: How much the inner disk radius is shifted (positive is to the right)

floor: The floor of the inner disk 0 means the inner intensity is zero and 1 means it is a large disk.

Comrade.amplitude — Method

amplitude(model, args)

Computes the visibility amplitude of model m at u,v positions u,v

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.amplitude — Method

amplitude(D)

Get the amplitude of a visibility datum

Comrade.amplitudes — Method

amplitudes(m, u, v)

Computes the amplitudes of the model m at the u,v positions u, v.

Notes

If this is a analytic model this is done lazily so the visibilites are only computed when accessed. Otherwise for numerical model computed with NFFT this is eager.

Comrade.arrayconfig — Method

arrayconfig(vis)

Extract the array configuration from a visibility EHT observation.

Comrade.basemodel — Method

basemodel(model)

Returns the base model from a modified model.

Comrade.bispectra — Method

bispectra(m, u1, v1, u2, v2, u3, v3)

Computes the bispectra of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the bispectra is only computed when accessed.

Comrade.bispectrum — Method

bispectrum(model, u1, v1, u2, v2, u3, v3)

Computes the complex bispectrum of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.closure_phase — Method

closure_phase(model, u1, v1, u2, v2, u3, v3)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.closure_phase — Method

closure_phase(D1, D2, D3)

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

Comrade.closure_phases — Method

closure_phases(m, u1, v1, u2, v2, u3, v3)

Computes the closure phases of the model m at the triangles (u1,v1), (u2,v2), (u3,v3).

Note this is done lazily so the closure_phases is only computed when accessed.

Comrade.coherencymatrix — Method

coherencymatrix(pimg, u, v)

Computes the coherency matrix of the polarized model.

Comrade.components — Method

components(m)

Returns the components for a composite model. This will return a Tuple with all the models you have constructed.

Comrade.convolved — Method

convolved(m1, m2)

Convolves two models m1 and m2. This is done lazily.

Comrade.create_cache — Function

create_cache(alg, img)

Creates a Fourier transform cache for a img. This is usually called internally for ModelImage. However, in certain circumstances the end-user can call this to speed up a method.

Comrade.create_cache — Method

create_cache(alg, img)

Creates the model cache given for the algorithm alg using the model and a image cache image

Comrade.extract_amp — Method

extract_amp(obs)

Extracts the visibility amplitudes from an ehtim observation object.

Any valid keyword arguments to add_amp in ehtim can be passed through extract_amp.

Returns an EHTObservation with visibility amplitude data

Comrade.extract_cphase — Method

extract_cphase(obs)

Extracts the closure phases from an ehtim observation object

Any valid keyword arguments to add_cphase in ehtim can be passed through extract_cphase.

Returns an EHTObservation with closure phases datums

Keyword arguments:

count: How the closures are formed, the available options are "min-correct", "min", "max"

cut_trivial: Cut the trivial triangles from the closures

uvmin: The flag to decide what are trivial triangles. Any baseline with ||(u,v)|| < uvmin are removed.

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_lcamp — Method

extract_lcamp(obs)

Extracts the log-closure amp. from an ehtim observation object

Any valid keyword arguments to add_logcamp in ehtim can be passed through extract_lcamp.

Returns an EHTObservation with log-closure amp. datums

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_vis — Method

extract_vis(obs)

Extracts the complex visibilities from an ehtim observation object

This grabs the raw data object from the obs object. Any keyword arguments are ignored.

Returns an EHTObservation with complex visibility data

Comrade.getuv — Method

getuv(ac)

Get the u, v positions of the array.

Comrade.load_ehtim — Method

load_ehtim()

Loads the eht-imaging library and stores it in the ehtim variable.

Notes

This will fail if ehtim isn't installed in the python installation that PyCall references.

Comrade.load_tpy — Method

load_tpy(file)

Load a ThemisPy style ascii EHT observation file.

Comrade.logclosure_amplitude — Method

logclosure_amplitude(model, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v3 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.logclosure_amplitudes — Method

logclosure_amplitudes(m, u1, v1, u2, v2, u3, v3, u4, v4)

Computes the log closure amplitudes of the model m at the quadrangles (u1,v1), (u2,v2), (u3,v3), (u4, v4).

Note this is done lazily so the log closure amplitude is only computed when accessed.

Comrade.modelimage — Method

Construct a ModelImage where just the model m is specified

Notes

If m IsAnalytic() is the visibility domain this is a no-op and just returns the model itself. Otherwise modelimage will guess a reasonable field of view based on the radialextent function. One can optionally pass the number of pixels nx and ny in each direction.

Comrade.modelimage — Method

modelimage(model, image)
modelimage(model, image, alg)
modelimage(model, image, alg, executor)
modelimage(m; fovx, fovy, nx, ny, pulse, alg, executor)

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg

Notes

For analytic models this is a no-op and just return the model. For non-analytic models this wraps the model in a object with an image and precomputes the fourier transform using alg.

Comrade.prior_sample — Method

prior_sample(post, args)

Samples the prior distribution from the posterior nsamples times.

Returns a Vector{NamedTuple} that can be used to initialize optimization and sample algorithms

Comrade.rad2μas — Method

rad2μas(x)

Converts a number from radians to μas

Comrade.renormed — Method

renormed(model, f)

Renormalizes the model m to have total flux flux.

Comrade.rescaleuv! — Method

rescaleuv!(data)

rescale the u-v lengths according to scale. This can be useful when you want the spatial scales to be in 1/μas instead of 1/rad

Comrade.rotated — Method

rotated(model, ξ)

Rotates the model by an amount ξ in radians.

Comrade.scantable — Method

`scantable(obs)`

Reorganizes the observation into a table of scans, where scan are defined by unique timestamps. To access the data you can use scalar indexing

Example

st = scantable(obs)
# Grab the first scan
scan1 = st[1]

# Acess the detections in the scan
scan1[1]

# grab e.g. the baselines
scan1[:baseline]

Comrade.shifted — Method

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy.

Comrade.smoothed — Method

smoothed(m, σ)

Smooths a model m with a Gaussian kernel with standard deviation σ.

Notes

This will created a convolved model under the hood.

Comrade.stations — Method

Get all the stations in a observation. The result is a vector of symbols.

Comrade.stretched — Method

stretched(model, α, β)

Stretches the model m according to the formula

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

Comrade.update_cache — Function

update_cache(cache, img)

Update the Fourier transform cache. This will reuse an FFT/NFFT plan saving some computational time.

Comrade.uviterator — Method

uviterator(dx, dy, nnx, nny)

Construct the u,v iterators for the Fourier transform of the image with pixel sizes dx, dy and number of pixels nx, ny

If you are extending Fourier transform stuff please use these functions to ensure that the centroid is being properly computed.

Comrade.uvpositions — Method

uvpositions(datum)

Get the uvp positions of an inferometric datum.

Comrade.uvtimefreq — Method

uvtimefreq(ac)

Get the u, v, time, freq of the array as a tuple.

Comrade.visibility — Method

visibility(pimg, u, v)

Computes the visibility in the stokes basis of the polarized model

Comrade.visibility — Method

Return the complex visibility of the visibility datum

Comrade.visibility — Method

visibility(mimg, args)

Computes the complex visibility of model m at u,v positions u,v

If you want to compute the visibilities at a large number of positions consider using the visibilities function which uses MappedArrays to compute a lazy, no allocation vector.

Comrade.μas2rad — Method

μas2rad(x)

Converts a number from μas to rad

ComradeBase.evpa — Method

evpa(pimg, u, v)

electric vector position angle or EVPA of the polarized model

ComradeBase.intensitymap! — Method

intensitymap!(pimg, pmodel)

Finds the polarized intensity map of the polarized model pmodel.

ComradeBase.m̆ — Method

m̆(pimg, u, v)

Computes the fractional linear polarization in the visibility domain

\[ \breve{m} = \frac{Q + iU}{I}\]

HypercubeTransform.ascube — Method

ascube(post::Posterior)

Construct a flattened version of the posterior, where the parameters are transformed to live in the unit hypercube. In astronomy parlance, we are transforming the variables to the unit hypercube. This is done using the HypercubeTransform package.

The transformed posterior can then be evaluated by the logdensityof(transformed_posterior,x) method following the DensityInterface, where x vector that lives in the unit hypercube. Note this already includes the jacobian of the transformation so this does not need to be added.

This transform is useful for NestedSampling methods that often assume that the model is written to live in the unit hypercube.

HypercubeTransform.asflat — Method

asflat(post::Posterior)

Construct a flattened version of the posterior, where the parameters are transformed so that their support is from (-∞, ∞). This uses TransformVariables

The transformed posterior can then be evaluated by the logdensityof(transformed_posterior,x) method following the DensityInterface, where x is a flattened vector of the infinite support variables. Note this already includes the jacobian of the transformation so this does not need to be added.

This is useful for optimization and sampling algorithms such as HMC that will use gradients to explore the posterior surface.

ParameterHandling.flatten — Method

flatten(post::Posterior)

Flatten the representation of the posterior. Internally this uses ParameterHandling to construct a flattened version of the posterior.

Note this is distinct from asflat that transforms the variables to live in (-∞,∞). Instead this method just flattens the repsentation of the model from a NamedTuple to a vector. This allows the easier integration to optimization and sampling algorithms.

RecipesBase.apply_recipe — Method

plot(image::AbstractModel)

where image is templated off of EHTImage struct.

Details

This was created to be close to the ehtim display object. It takes an EHTImage object and plots it according to EHT conventions.

Note that is does not save the figure.

RecipesBase.apply_recipe — Method

plot(image::IntensityMap)

where image is templated off of EHTImage struct.

Details

This was created to be close to the ehtim display object. It takes an EHTImage object and plots it according to EHT conventions.

Note that is does not save the figure.

TransformVariables.inverse — Method

inverse(posterior::TransformedPosterior, x)

Transforms the value model parameters x into the flattened transformed space.

TransformVariables.transform — Method

transform(posterior::TransformedPosterior, x)

Transforms the value x into parameter space, i.e. usually a NamedTuple.

Comrade.AbstractCache — Type

abstract type AbstractCache

This defines an abstract cache that can be used to hold or precompute some computations.

Comrade.AbstractInterferometryDatum — Type

abstract type AbstractInterferometryDatum{T}

An abstract type for all VLBI interfermetry data types. See EHTVisibilityDatum for an example.

Comrade.AbstractModifier — Type

abstract type AbstractModifier{M<:ComradeBase.AbstractModel} <: ComradeBase.AbstractModel

Abstract type for image modifiers. These are some model wrappers that can transform any model using simple Fourier transform properties. To see the implemented modifier

Comrade.AddModel — Type

struct AddModel{T1, T2} <: Comrade.CompositeModel{T1, T2}

Adds two models together to create composite models. Note that I may change this in the future so make it easier on the compiler, i.e. make the composite model a fancy model vector and heap allocate stuff. This should help when combining multiple models together.

Comrade.ArrayBaselineDatum — Type

struct ArrayBaselineDatum{T}

A single datum of an observing array.

Comrade.ArrayConfiguration — Type

abstract type ArrayConfiguration

This defined the abstract type for an array configuration. Namely, baseline times, SEFD's, bandwidth, observation frequencies, etc.

Comrade.ClosureConfig — Type

struct ClosureConfig{A, D} <: Comrade.ArrayConfiguration

Array config file for closure quantities. This stores the design matrix designmat that transforms from visibilties to closure products.

Comrade.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ComradeBase.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

ComradeBase.intensity_point if model intensity is IsAnalytic

intensitymap! if model intensity is NotAnalytic

intensitymap if model intensity is NotAnalytic

flux

Comrade.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: Comrade.GeometricModel

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

router

Outer radius of the crescent

rinner

Inner radius of the crescent (i.e. inside this radius there is a hole)

shift

Displacment of the inner disk radius

slash

Strength of the linear slash. Note that s∈[0.0,1.0] to ensure positivity in the image.

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using Comrade.jl's model composition functionality.

Comrade.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: Comrade.CompositeModel{M1, M2}

Convolves two models m1 and m2.

Notes

This is the non-exported constructor. The user should call the convolved function during use

Comrade.DFTAlg — Type

DFTAlg

Uses a discrete fourier transform. This is not very efficient for larger images. In those cases NFFTAlg or FFTAlg are more reasonable. For small images this is a reasonable choice especially since it's easy to define derivatives.

Comrade.EHTArrayConfiguration — Type

struct EHTArrayConfiguration{F, T<:AbstractArray} <: Comrade.ArrayConfiguration

Stores all the non-visibility data products for an EHT array. This is useful when evaluating model visibilities.

frequency

Observing frequency (Hz)

bandwidth

Observing bandwith (Hz)

data

A struct array of ArrayBaselineDatum holding time, u, v, baselines.

Comrade.EHTClosurePhaseDatum — Type

struct EHTClosurePhaseDatum{T<:Number} <: Comrade.ClosureProducts{T<:Number}

A Datum for a single closure phase.

phase

closure phase (rad)

error

error of the closure phase assuming the high-snr limit

u1

u (λ) of first station

v1

v (λ) of first station

u2

u (λ) of second station

v2

v (λ) of second station

u3

u (λ) of third station

v3

v (λ) of third station

time

Measured time of closure phase in hours

triangle

station baselines used

Comrade.EHTLogClosureAmplitudeDatum — Type

struct EHTLogClosureAmplitudeDatum{T<:Number} <: Comrade.ClosureProducts{T<:Number}

A Datum for a single log closure amplitude.

amp

log-closure amplitude

error

log-closure amplitude error in the high-snr limit

u1

u (λ) of first station

v1

v (λ) of first station

u2

u (λ) of second station

v2

v (λ) of second station

u3

u (λ) of third station

v3

v (λ) of third station

u4

u (λ) of fourth station

v4

v (λ) of fourth station

time

Observation time of the quadrangle

quadrangle

station codes for the quadrangle

Comrade.EHTObservation — Type

The main data product type in Comrade this stores the data which can be a StructArray of any AbstractInterferometryDatum type.

(

data

StructArray of data productts

config

Array config holds ancillary information about array

mjd

modified julia date of the observation

ra

RA of the observation in J2000 (deg)

dec

DEC of the observation in J2000 (deg)

bandwidth

bandwidth of the observation (Hz)

frequency

frequency of the observation (Hz)

source

Common source name

timetype

Time zone used.

)

Comrade.EHTVisibilityAmplitudeDatum — Type

A struct holding the information for a single measured visibility amplitude.

amp

amplitude (Jy)

error

error of the visibility amplitude (Jy)

u

x-direction baseline length in λ

v

y-direction baseline length in λ

time

Time of the observation in hours

baseline

station baseline codes

Comrade.EHTVisibilityDatum — Type

A struct holding the information for a single measured visibility.

visr

real component of the visibility (Jy)

visi

imaginary component of the visibility (Jy)

error

error of the visibility (Jy)

u

x-direction baseline length in λ

v

y-direction baseline length in λ

time

Time of the observation in hours

baseline

station baseline codes

Comrade.ExtendedRing — Type

struct ExtendedRing{F} <: Comrade.GeometricModel

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

Note

We mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

shape

shape of the radial distribution

Comrade.FFTAlg — Type

struct FFTAlg <: Comrade.FourierTransform

Fourier transform type that specifies we will use the FFTW package to compute the Fourier transform.

Fields

padfac

The amount to pad the image by. Note we actually round up to the nearest factor of 2, but this will be improved in the future to use small primes

Comrade.FFTCache — Type

struct FFTCache{A<:FFTAlg, P, I, S} <: Comrade.AbstractCache

The cache used when the FFT algorithm is used to compute visibilties. This is an internal type and is not part of the public API

Fields

alg

plan

FFTW Plan

img

image cache

sitp

FFT interpolator function

Comrade.GainModel — Type

`GainModel`

Construct the corruption model for the telescope gains

Comrade.Gaussian — Type

struct Gaussian{T} <: Comrade.GeometricModel

Gaussian geometrical model. This is a Gaussian with unit flux and standard deviation.

Notes

To change the Gaussian flux, and shape please use the modifier functions

Comrade.GeometricModel — Type

abstract type GeometricModel <: ComradeBase.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain.

Comrade.MRing — Type

struct MRing{T, N} <: Comrade.GeometricModel

m-ring geometric model. This corresponds to a delta ring with a fourier expansion in θ. The m in m-ring refers to the order of the Fourier expansion. The radius is unity.

Comrade.ModelImage — Type

struct ModelImage{M, I, C} <: Comrade.AbstractModelImage{M}

Container for non-analytic model that contains a image cache which will hold the image, a Fourier transform cache C, which usually an instance of a <: FourierCache which usually holds a interpolator that allows you to compute visibilities.

Notes

This is an internal implementation detail that shouldn't usually be called directly. Instead the user should use the exported function modelimage, for example

using Comrade
m = ExtendedRing(20.0, 5.0)

# This creates an version where the image is dynamically specified according to the
# radial extent of the image
mimg = modelimage(m) # you can also optionally pass the number of pixels nx and ny

# Or you can create an IntensityMap
img = intensitymap(m, 100.0, 100.0, 512, 512)
mimg = modelimage(m, img)

Comrade.MultiRadioLikelihood — Method

`MultiRadioLikelihood(lklhd1, lklhd2, ...)`

Combines multiple likelihoods into one object that is useful for fitting multiple days/frequencies.

Comrade.NFFTAlg — Type

NFFTAlg

Uses a non-uniform FFT to compute the visibilities. You can optionally pass uv which are the uv positions you will compute the NFFT at. This can allow for the NFFT plan to be cached improving performance

Fields

padfac

Amount to pad the image

m

Controls the accuracy of the NFFT usually don't need to change this

Comrade.NonAnalyticTest — Type

`NonAnalyticTest`

An internal model used primarly for testing. Any model passed to it will be interpreted as not having an analytic Fourier transform.

Comrade.ParabolicSegment — Type

struct ParabolicSegment{F} <: Comrade.GeometricModel

A delta parabolic segment in the image domain. The segment is centered at zero, with roots ±1 and a yintercept of 1.

Comrade.ParabolicSegment — Method

ParabolicSegment(a, h)

A parabolic segment with x-intercepts ±aand a yintercept ofh``.

Note

This is just a convenience function for stretched(ParabolicSegment(), a, h)

Comrade.PolarizedModel — Type

struct PolarizedModel{TI, TQ, TU, TV} <: ComradeBase.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

Comrade.Posterior — Type

Posterior(lklhd, prior, model)

Posterior density that follows obeys the DensityInferface

The expected arguments are:

lklhd: Which should be an intance of RadioLikelihood with whatever data products you want to fit

prior: This should be a NamedTuple with the priors for each model ParameterHandling

model: Function that takes a NamedTuple of parameters and constructs the Comrade model.

To evaluate the logdensity of the posterior use can either use DensityInterface.logdensityof or logdensity.

Comrade.RadioLikelihood — Method

`RadioLikelihood(data1, data2, ...)`

Forms a radio likelihood from a set of data products. These data products must share the same array data/configuration. If you want to form a likelihood from multiple arrays such as when fitting different wavelengths or days, you can combine them using MultiRadioLikelihood

lklhd1 = RadioLikelihood(dcphase1, dlcamp1)
lklhd2 = RadioLikelihood(dcphase2, dlcamp2)

lklhd = MultiRadioLikelihood(lklhd1, lklhd2)

Comrade.RenormalizedModel — Type

struct RenormalizedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Renormalizes the flux of the model to the new value flux. We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

Comrade.Ring — Type

struct Ring{T} <: Comrade.GeometricModel

m-ring geometric model. This corresponds to a delta ring with a fourier expansion in θ. The m in m-ring refers to the order of the Fourier expansion. The radius is unity.

Comrade.RotatedModel — Type

struct RotatedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Type for the rotated model. This is more fine grained constrol of rotated model. In most use cases the end-user should be using the rotate method e.g.

rotate(model, ξ)

Comrade.ShiftedModel — Type

struct ShiftedModel{T, M<:ComradeBase.AbstractModel} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Shifts the model by Δx units in the x-direction and Δy units in the y-direction.

Comrade.StretchedModel — Type

struct StretchedModel{M<:ComradeBase.AbstractModel, T} <: Comrade.AbstractModifier{M<:ComradeBase.AbstractModel}

Stretched the model in the x and y directions, i.e. the new intensity is

\[ I_s(x,y) = 1/(αβ) I(x/α, y/β),\]

where were renormalize the intensity to preserve the models flux.

Comrade.TransformedPosterior — Type

`

struct TransformedPosterior{P<:Posterior, T}

A transformed version of the posteriorlpost`. This is an internal type that an end user shouldn't have to directly construct.

To construct a transformed posterior see the asflat, ascube, and flatten functions.

When evaluating the logdensity of the TransformedPosterior, logdenisty will include any jacobian terms automatically, and expects the argument to be the vector of transformed quantities. This is usually easier to incorporate with samplers, and optimizers, which expect homogenous vectors as arguments.