Comrade API

Contents

Comrade API

Contents

Index

Model Definitions

Data Types

eht-imaging interface

Bayesian Tools

Misc

Internal (Not Public API)

Index

Comrade.AbstractInterferometryDatum

Comrade.AddModel

Comrade.ArrayBaselineDatum

Comrade.ArrayConfiguration

Comrade.BSplinePulse

Comrade.BicubicPulse

Comrade.Butterworth

Comrade.CalPrior

Comrade.CalPrior

Comrade.CalTable

Comrade.ClosureConfig

Comrade.CompositeModel

Comrade.ConcordanceCrescent

Comrade.ContinuousImage

Comrade.ConvolvedModel

Comrade.DFTAlg

Comrade.DFTAlg

Comrade.DFTAlg

Comrade.DFTAlg

Comrade.DeltaPulse

Comrade.DesignMatrix

Comrade.Disk

Comrade.EHTArrayConfiguration

Comrade.EHTClosurePhaseDatum

Comrade.EHTCoherencyDatum

Comrade.EHTLogClosureAmplitudeDatum

Comrade.EHTObservation

Comrade.EHTVisibilityAmplitudeDatum

Comrade.EHTVisibilityDatum

Comrade.ExtendedRing

Comrade.FFTAlg

Comrade.FFTCache

Comrade.FixedSeg

Comrade.Gaussian

Comrade.GeometricModel

Comrade.IntegSeg

Comrade.IsCube

Comrade.IsFlat

Comrade.JonesCache

Comrade.JonesModel

Comrade.JonesPairs

Comrade.MRing

Comrade.ModelImage

Comrade.MultiRadioLikelihood

Comrade.NFFTAlg

Comrade.NUFTCache

Comrade.NonAnalyticTest

Comrade.ObsSegmentation

Comrade.ObservedNUFT

Comrade.ParabolicSegment

Comrade.PolarizedModel

Comrade.Posterior

Comrade.Pulse

Comrade.RIMEModel

Comrade.RadioLikelihood

Comrade.RaisedCosinePulse

Comrade.Ring

Comrade.Scan

Comrade.ScanSeg

Comrade.ScanTable

Comrade.SlashedDisk

Comrade.ThreadedModel

Comrade.TrackSeg

Comrade.TransformCache

Comrade.TransformedPosterior

Comrade.ZeroModel

Base.:+

Comrade.Crescent

Comrade.PoincareSphere2Map

Comrade.added

Comrade.amplitude

Comrade.amplitude

Comrade.amplitude

Comrade.amplitudes

Comrade.arrayconfig

Comrade.baselines

Comrade.basemodel

Comrade.bispectra

Comrade.bispectrum

Comrade.caltable

Comrade.caltable

Comrade.caltable

Comrade.closure_phase

Comrade.closure_phase

Comrade.closure_phases

Comrade.components

Comrade.convolved

Comrade.corrupt

Comrade.create_cache

Comrade.extract_FRs

Comrade.extract_amp

Comrade.extract_coherency

Comrade.extract_cphase

Comrade.extract_lcamp

Comrade.extract_vis

Comrade.fouriermap

Comrade.getdata

Comrade.getuv

Comrade.getuvtimefreq

Comrade.jonesD

Comrade.jonesG

Comrade.jonesStokes

Comrade.jonesT

Comrade.likelihood

Comrade.load

Comrade.load_ehtim

Comrade.load_ehtim_uvfits

Comrade.logclosure_amplitude

Comrade.logclosure_amplitudes

Comrade.modelimage

Comrade.posangle

Comrade.prior_sample

Comrade.rad2μas

Comrade.renormed

Comrade.rotated

Comrade.save

Comrade.scale_image

Comrade.scale_uv

Comrade.scan_average

Comrade.scantable

Comrade.shifted

Comrade.smoothed

Comrade.stations

Comrade.stretched

Comrade.transform_image

Comrade.transform_uv

Comrade.unmodified

Comrade.update_cache

Comrade.uviterator

Comrade.uvpositions

Comrade.visibility

Comrade.μas2rad

ComradeBase.evpa

ComradeBase.mbreve

HypercubeTransform.ascube

HypercubeTransform.asflat

ParameterHandling.flatten

StatsBase.sample

TransformVariables.inverse

TransformVariables.transform

Model Definitions

Calibration Models

Comrade.corrupt — Function

corrupt(vis, j1, j2)

Corrupts the model coherency matrices with the Jones matrices j1 for station 1 and j2 for station 2.

Comrade.CalTable — Type

struct CalTable{T, G<:(AbstractVecOrMat)}

A Tabes of calibration quantities. The columns of the table are the telescope station codes. The rows are the calibration quantities at a specific time stamp. This user should not use this struct directly. Instead that should call caltable.

Comrade.caltable — Method

caltable(g::JonesCache, jterms::AbstractVector)

Convert the JonesCache g and recovered Jones/corruption elements jterms into a CalTable which satisfies the Tables.jl interface.

Example

ct = caltable(gcache, gains)

# Access a particular station (here ALMA)
ct[:AA]
ct.AA

# Access a the first row
ct[1, :]

Comrade.caltable — Method

caltable(obs::EHTObservation, gains::AbstractVector)

Create a calibration table for the observations obs with gains. This returns a CalTable object that satisfies the Tables.jl interface. This table is very similar to the DataFrames interface.

Example

ct = caltable(obs, gains)

# Access a particular station (here ALMA)
ct[:AA]
ct.AA

# Access a the first row
ct[1, :]

Comrade.DesignMatrix — Type

struct DesignMatrix{X, M<:AbstractArray{X, 2}, T, S} <: AbstractArray{X, 2}

Internal type that holds the gain design matrices for visibility corruption.

Comrade.JonesCache — Type

struct JonesCache{D1, D2, S, Sc} <: Comrade.AbstractJonesCache

Holds the ancillary information for a the design matrix cache for Jones matrices. That is, it defines the cached map that moves from model visibilities to the corrupted voltages that are measured from the telescope.

Fields

m1: Design matrix for the first station

m2: Design matrix for the second station

seg: Segmentation schemes for this cache

schema: Gain Schema

Comrade.TransformCache — Type

struct TransformCache{M, B<:PolBasis} <: Comrade.AbstractJonesCache

Holds various transformations that move from the measured telescope basis to the chosen on sky reference basis.

Fields

T1: Transform matrices for the first stations

T2: Transform matrices for the second stations

refbasis: Reference polarization basis

Comrade.JonesModel — Type

JonesModel(jones::JonesPairs, model, tcache::TransformCache)
JonesModel(jones::JonesPairs, model, refbasis::PolBasis=CirBasis())

Constructs a JonesModel from a jones pairs that describe the intrument model and the model which describes the on-sky polarized visibilities. The third argument can either be the tcache that converts from the model coherency basis to the instrumental basis, or just the refbasis that will be used when constructing the model coherency matrices.

Comrade.CalPrior — Type

CalPrior(dists, cache::JonesCache, reference=:none)

Creates a distribution for the gain priors for gain cache cache. The dists should be a NamedTuple of Distributions, where each name corresponds to a telescope or station in the observation. The resulting type is a subtype of the Distributions.AbstractDistribution so the usual Distributions interface should work.

Example

For the 2017 observations of M87 a common CalPrior call is:

julia> gdist = CalPrior((AA = LogNormal(0.0, 0.1),
                   AP = LogNormal(0.0, 0.1),
                   JC = LogNormal(0.0, 0.1),
                   SM = LogNormal(0.0, 0.1),
                   AZ = LogNormal(0.0, 0.1),
                   LM = LogNormal(0.0, 1.0),
                   PV = LogNormal(0.0, 0.1)
                ), cache)

julia> x = rand(gdist)
julia> logdensityof(gdist, x)

CalPrior(dist0::NamedTuple, dist_transition::NamedTuple, jcache::SegmentedJonesCache)

Constructs a calibration prior in two steps. The first two arguments have to be a named tuple of distributions, where each name corresponds to a site. The first argument is gain prior for the first time stamp. The second argument is the segmented gain prior for each subsequent time stamp. For instance, if we have

dist0 = (AA = Normal(0.0, 1.0), )
distt = (AA = Normal(0.0, 0.1), )

then the gain prior for first time stamp that AA obserserves will be Normal(0.0, 1.0). The next time stamp gain is the construted from

g2 = g1 + ϵ1

where ϵ1 ~ Normal(0.0, 0.1) = distt.AA, and g1 is the gain from the first time stamp. In other words distt is the uncorrelated transition probability when moving from timestamp i to timestamp i+1. For the typical pre-calibrated dataset the gain prior on distt can be tighter than the prior on dist0.

Comrade.CalPrior — Method

CalPrior(dists, cache::JonesCache, reference=:none)

Creates a distribution for the gain priors for gain cache cache. The dists should be a NamedTuple of Distributions, where each name corresponds to a telescope or station in the observation. The resulting type is a subtype of the Distributions.AbstractDistribution so the usual Distributions interface should work.

Example

For the 2017 observations of M87 a common CalPrior call is:

julia> gdist = CalPrior((AA = LogNormal(0.0, 0.1),
                   AP = LogNormal(0.0, 0.1),
                   JC = LogNormal(0.0, 0.1),
                   SM = LogNormal(0.0, 0.1),
                   AZ = LogNormal(0.0, 0.1),
                   LM = LogNormal(0.0, 1.0),
                   PV = LogNormal(0.0, 0.1)
                ), cache)

julia> x = rand(gdist)
julia> logdensityof(gdist, x)

Missing docstring.

Missing docstring for Comrade.CalPrior(::NamedTuple, ::NamedTuple, ::Comrade.SegmentedJonesCache). Check Documenter's build log for details.

Comrade.RIMEModel — Type

abstract type RIMEModel <: ComradeBase.AbstractModel

Abstract type that encompasses all RIME style corruptions.

Comrade.ObsSegmentation — Type

abstract type ObsSegmentation

The data segmentation scheme to use. This is important for constructing a JonesCache

Comrade.IntegSeg — Type

struct IntegSeg{S} <: Comrade.ObsSegmentation

Data segmentation such that the quantity is constant over a correlation integration.

Comrade.ScanSeg — Type

struct ScanSeg{S} <: Comrade.ObsSegmentation

Data segmentation such that the quantity is constant over a scan.

Warning

Currently we do not explicity track the telescope scans. This will be fixed in a future version. Right now ScanSeg and TrackSeg are the same

Comrade.TrackSeg — Type

struct TrackSeg <: Comrade.ObsSegmentation

Data segmentation such that the quantity is constant over a track, i.e., the observation "night".

Comrade.FixedSeg — Type

struct FixedSeg{T} <: Comrade.ObsSegmentation

Enforces that the station calibraton value will have a fixed value. This is most commonly used when enforcing a reference station for gain phases.

Missing docstring.

Missing docstring for Comrade.jonescache(::Comrade.EHTObservation, ::NamedTuple). Check Documenter's build log for details.

Comrade.jonesStokes — Function

jonesStokes(g1::AbstractArray, gcache::AbstractJonesCache)
jonesStokes(f, g1::AbstractArray, gcache::AbstractJonesCache)

Construct the Jones Pairs for the stokes I image only. That is, we only need to pass a single vector corresponding to the gain for the stokes I visibility. This is for when you only want to image Stokes I. The first argument is optional and denotes a function that is applied to every element of jones cache. For instance if g1 and g2 are the log-gains then f=exp will convert them into the gains.

Warning

In the future this functionality may be removed when stokes I fitting is replaced with the more correct trace(coherency), i.e. RR+LL for a circular basis.

Comrade.jonesG — Function

jonesG(g1::AbstractVector, g2::AbstractVector, jcache::AbstractJonesCache)
jonesG(f, g1::AbstractVector, g2::AbstractVector, jcache::AbstractJonesCache)

Constructs the pairs Jones G matrices for each pair of stations. The g1 are the gains for the first polarization basis and g2 are the gains for the other polarization. The first argument is optional and denotes a function that is applied to every element of jones cache. For instance if g1 and g2 are the log-gains then f=exp will convert them into the gains.

The layout for each matrix is as follows:

    g1 0
    0  g2

Comrade.jonesD — Function

jonesD(d1::AbstractVector, d2::AbstractVector, jcache::AbstractJonesCache)
jonesD(f, d1::AbstractVector, d2::AbstractVector, jcache::AbstractJonesCache)

Constructs the pairs Jones D matrices for each pair of stations. The d1 are the d-termsfor the first polarization basis and d2 are the d-terms for the other polarization. The first argument is optional and denotes a function that is applied to every element of jones cache. For instance if d1 and d2 are the log-dterms then f=exp will convert them into the dterms.

The layout for each matrix is as follows:

    1  d1
    d2 1

Comrade.jonesT — Function

jonesT(tcache::TransformCache)

Returns a JonesPair of matrices that transform from the model coherency matrices basis to the on-sky coherency basis, this includes the feed rotation and choice of polarization feeds.

Comrade.PoincareSphere2Map — Function

PoincareSphere2Map(I, p, X, grid)
PoincareSphere2Map(I::IntensityMap, p, X)

Constructs an polarized intensity map model using the Poincare parameterization. The arguments are:

I is a grid of fluxes for each pixel.

p is a grid of numbers between 0, 1 and the represent the total fractional polarization

X is a grid, where each element is 3 numbers that represents the point on the Poincare sphere that is, X[1,1] is a NTuple{3} such that ||X[1,1]|| == 1.

grid is the dimensional grid that gives the pixels locations of the intensity map.

Note

If I is an IntensityMap then grid is not required since the same grid that was use for I will be used to construct the polarized intensity map

Warning

The return type for this function is a polarized image object, however what we return is not considered to be part of the stable API so it may change suddenly.

Comrade.caltable — Function

caltable(args...)

Creates a calibration table from a set of arguments. The specific arguments depend on what calibration you are applying.

Comrade.JonesPairs — Type

struct JonesPairs{T, M1<:AbstractArray{T, 1}, M2<:AbstractArray{T, 1}}

Holds the pairs of Jones matrices for the first and second station of a baseline.

Fields

m1: Vector of jones matrices for station 1

m2: Vector of jones matrices for station 2

Combinators

Base.:+ — Method

Base.:+(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(m1+m2, 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

Comrade.added — Function

added(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(added(m1,m2), 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

Comrade.convolved — Function

convolved(m1::AbstractModel, m2::AbstractModel)

Convolve two models to create a composite ConvolvedModel.

julia> m1 = Ring()
julia> m2 = Disk()
julia> convolved(m1, m2)

Comrade.components — Function

components(m::AbstractModel)

Returns the model components for a composite model. This will return a Tuple with all the models you have constructed.

Example

julia> m = Gaussian() + Disk()
julia> components(m)
(Gaussian{Float64}(), Disk{Float64}())

Comrade.smoothed — Function

smoothed(m::AbstractModel, σ::Number)

Smooths a model m with a Gaussian kernel with standard deviation σ.

Notes

This uses convolved to created the model, i.e.

julia> m1 = Disk()
julia> m2 = Gaussian()
julia> convolved(m1, m2) == smoothed(m1, 1.0)
true

Comrade.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ComradeBase.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

ComradeBase.intensity_point if model intensity is IsAnalytic

intensitymap_numeric! if model intensity is NotAnalytic

intensitymap_numeric if model intensity is NotAnalytic

flux

radialextent

In addition there are additional optional methods a person can define if needed:

intensitymap_analytic! if model intensity is IsAnalytic (optional)

intensitymap_analytic if model intensity is IsAnalytic (optional)

visibilities_analytic if visanalytic is IsAnalytic (optional)

visibilities_numeric if visanalytic is Not Analytic (optional)

visibilities_analytic! if visanalytic is IsAnalytic (optional)

visibilities_numeric! if visanalytic is Not Analytic (optional)

Comrade.AddModel — Type

struct AddModel{T1, T2} <: Comrade.CompositeModel{T1, T2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call added or Base.+ when constructing a model

Example

julia> m1 = Disk() + Gaussian()
julia> m2 = added(Disk(), Gaussian()) + Ring()

Comrade.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: Comrade.CompositeModel{M1, M2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call convolved. Also see smoothed(m, σ) for a simplified function that convolves a model m with a Gaussian with standard deviation σ.

Geometric and Image Models

Comrade.GeometricModel — Type

abstract type GeometricModel <: ComradeBase.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain. As a result a user only needs to implement the following methods

visibility_point

intensity_point

radialextent

Note that if the geometric model isn't analytic then the usual methods listed in Comrade.AbstractModel for non-analytic models need to be implemented.

Comrade.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: Comrade.GeometricModel

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash and shadow depth. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

router: Outer radius of the crescent

rinner: Inner radius of the crescent (i.e. inside this radius there is a hole)

shift: Displacment of the inner disk radius

slash: Strength of the linear slash. Note that s∈[0.0,1.0] to ensure positivity in the image.

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using Comrade.jl's model composition functionality.

Comrade.Crescent — Function

Creates a Kamruddin and Dexter crescent model. This works by composing two disk models together.

Arguments

router: The radius of the outer disk

rinner: The radius of the inner disk

shift: How much the inner disk radius is shifted (positive is to the right)

floor: The floor of the inner disk 0 means the inner intensity is zero and 1 means it is a large disk.

Comrade.Disk — Type

Disk{T}() where {T}

Tophat disk geometrical model, i.e. the intensity profile

\[ I(x,y) = \begin{cases} \pi^{-1} & x^2+y^2 < 1 \\ 0 & x^2+y^2 \geq 0 \end{cases}\]

i.e. a unit radius and unit flux disk.

By default if T isn't given, Disk defaults to Float64

Comrade.SlashedDisk — Type

SlashedDisk{T}(slash::T) where {T}

Tophat disk geometrical model, i.e. the intensity profile

\[ I(x,y) = \begin{cases} \pi^{-1} & x^2+y^2 < 1 \\ 0 & x^2+y^2 \geq 0 \end{cases}\]

i.e. a unit radius and unit flux disk.

By default if T isn't given, Disk defaults to Float64

Comrade.ExtendedRing — Type

struct ExtendedRing{F<:Number} <: Comrade.GeometricModel

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

The formula in the image domain is given by

I(r,θ) = βᵅrᵅ⁻²exp(-β/r)/2πΓ(α)

where α = shape and β = shape+1

Note

We mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

shape: shape of the radial distribution

Note that if T isn't specified at construction then it defaults to Float64.

Comrade.Gaussian — Type

struct Gaussian{T} <: Comrade.GeometricModel

Gaussian with unit standard deviation and flux.

By default if T isn't given, Gaussian defaults to Float64

Comrade.MRing — Type

struct MRing{T, V<:Union{AbstractArray{T, 1}, Tuple{Vararg{T, N}} where {N, T}}} <: Comrade.GeometricModel

m-ring geometric model. This is a infinitely thin unit flux delta ring whose angular structure is given by a Fourier expansion. That is,

I(r,θ) = (2π)⁻¹δ(r-1)∑ₙ(αₙcos(nθ) - βₙsin(nθ))

The N in the type defines the order of the Fourier expansion.

Fields

α: Real Fourier mode coefficients

β: Imaginary Fourier mode coefficients

Comrade.Ring — Type

struct Ring{T} <: Comrade.GeometricModel

A infinitely thin ring model, whose expression in the image domain is I(r,θ) = δ(r - 1)/2π i.e. a unit radius and flux delta ring.

By default if T isn't given, Gaussian defaults to Float64

Comrade.ParabolicSegment — Type

struct ParabolicSegment{T} <: Comrade.GeometricModel

A infinitely thin parabolic segment in the image domain. The segment is centered at zero, with roots ±1 and a yintercept of 1.

Note that if T isn't specified at construction then it defaults to Float64.

Comrade.ContinuousImage — Type

ContinuousImage{A<:IntensityMap, P} <: AbstractModel
ContinuousImage(img::Intensitymap, kernel)

The basic continuous image model for Comrade. This expects a IntensityMap style object as its imag as well as a image kernel or pulse that allows you to evaluate the image at any image and visibility location. The image model is

I(x,y) = ∑ᵢ Iᵢⱼ κ(x-xᵢ, y-yᵢ)

where Iᵢⱼ are the flux densities of the image img and κ is the intensity function for the kernel.

Comrade.ZeroModel — Type

struct ZeroModel{T} <: ComradeBase.AbstractModel

Defines a model that is empty that is it return zero for everything.

Notes

This returns 0 by using FillArrays so everything should be non-allocating

Image Pulses

Comrade.Pulse — Type

Pulse

Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

Comrade.DeltaPulse — Type

struct DeltaPulse{T} <: Comrade.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

Comrade.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

Comrade.RaisedCosinePulse — Type

struct RaisedCosinePulse{T} <: Comrade.Pulse

RaisedCosinePulse()
RaisedCosinePulse(rolloff)

Raised cosine pulse function. This tends to be a very flat response, where the roll off controls the speed of decay. By default we set rolloff = 0.5.

Comrade.BicubicPulse — Type

struct BicubicPulse{T} <: Comrade.Pulse

BicubicPulse(b = 0.5)

The bicubic pulse for imaging. This pulse tends to have a flat spectrum but for most values of b can produce negative intensities in an image. This is the pulse used in Broderick et al. 2020.

Comrade.Butterworth — Type

Butterworth{N}()
Butterworth{N, T}()

Construct a model that corresponds to the Butterworth filter of order N. The type of the output is given by T and if not given defaults to Float64

Model Image (non analytic FFT)

Comrade.create_cache — Function

create_cache(alg::AbstractFourierTransform, img::AbstractIntensityMap)

Creates a Fourier transform cache for a img using algorithm alg. For non-analytic visibility models this can significantly speed up computations.

Examples

julia> u,v = rand(100), rand(100)
julia> cache = create_cache(DFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(NFFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(FFTAlg(), IntensityMap(randn(50,50), 10.0, 10.0))

Comrade.update_cache — Function

update_cache(cache, img)

Update the Fourier transform cache. This will reuse an FFT/NFFT plan saving some computational time.

Note

This is an intenal method than an end user shouldn't have to usually call.

Comrade.modelimage — Function

modelimage(model::AbstractModel, image::AbstractIntensityMap, alg=FFTAlg())

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg

Notes

For analytic models this is a no-op and returns the model. For non-analytic models this creates a ModelImage object which uses alg to compute the non-analytic Fourier transform.

modelimage(model, cache::AbstractCach))

Construct a ModelImage from the model and using a precompute Fourier transform cache. You can optionally specify th which will compute the internal image buffer using the`.

Example

julia> m = ExtendedRing(10.0)
julia> cache = create_cache(DFTAlg(), IntensityMap(zeros(128, 128), 50.0, 50.0)) # used threads to make the image
julia> mimg = modelimage(m, cache, true)

Notes

For analytic models this is a no-op and returns the model.

modelimage(m;
           fovx=2*radialextent(m),
           fovy=2*radialextent(m),
           nx=512,
           ny=512,
           alg=FFTAlg(),
           pulse=ComradeBase.DeltaPulse(),
            )

Construct a ModelImage where just the model m is specified.

If fovx or fovy aren't given modelimage will guess a reasonable field of view based on the radialextent function. nx and ny are the number of pixels in the x and y direction. The pulse is the pulse used for the image and alg

Notes

For analytic models this is a no-op and returns the model.

modelimage(img::ContinuousImage, alg=NFFTAlg())

Create a model image directly using an image, i.e. treating it as the model. You can optionally specify the Fourier transform algorithm using alg

modelimage(img::ContinuousImage, cache::AbstractCache)

Create a model image directly using an image, i.e. treating it as the model. Additionally reuse a previously compute image cache. This can be used when directly modeling an image of a fixed size and number of pixels.

Comrade.uviterator — Function

uviterator(nx, dx, ny dy)

Construct the u,v iterators for the Fourier transform of the image with pixel sizes dx, dy and number of pixels nx, ny

If you are extending Fourier transform stuff please use these functions to ensure that the centroid is being properly computed.

Comrade.fouriermap — Function

fouriermap(m, x)

Create a Fourier or visibility map of a model m where the image is specified in the image domain by the pixel locations x and y

Comrade.ModelImage — Type

struct ModelImage{M, I, C} <: Comrade.AbstractModelImage{M}

Container for non-analytic model that contains a image cache which will hold the image, a Fourier transform cache, which usually an instance of a <: FourierCache.

Note

This is an internal implementation detail that shouldn't usually be called directly. Instead the user should use the exported function modelimage, for example

using Comrade
m = ExtendedRing(20.0, 5.0)

# This creates an version where the image is dynamically specified according to the
# radial extent of the image
mimg = modelimage(m) # you can also optionally pass the number of pixels nx and ny

# Or you can create an IntensityMap
img = intensitymap(m, 100.0, 100.0, 512, 512)
mimg = modelimage(m, img)

# Or precompute a cache
cache = create_cache(FFTAlg(), img)
mimg = modelimage(m, cache)

Comrade.DFTAlg — Type

DFTAlg

Uses a discrete fourier transform. This is not very efficient for larger images. In those cases NFFTAlg or FFTAlg are more reasonable. For small images this is a reasonable choice especially since it's easy to define derivatives.

Comrade.DFTAlg — Method

DFTAlg(obs::EHTObservation)

Create an algorithm object using the direct Fourier transform object from the observation obs. This will extract the uv positions from the observation to allow for a more efficient FT cache.

Comrade.DFTAlg — Method

DFTAlg(u::AbstractArray, v::AbstractArray)

Create an algorithm object using the direct Fourier transform object using the uv positions u, v allowing for a more efficient transform.

Comrade.DFTAlg — Method

DFTAlg(ac::ArrayConfiguration)

Create an algorithm object using the direct Fourier transform object from the array configuration ac. This will extract the uv positions from the observation to allow for a more efficient FT cache.

Comrade.FFTAlg — Type

struct FFTAlg <: Comrade.FourierTransform

Use an FFT to compute the approximate numerical visibilities of a model. For a DTFT see DFTAlg or for an NFFT NFFTAlg

Fields

padfac: The amount to pad the image by. Note we actually round up to the nearest factor of 2, but this will be improved in the future to use small primes

Comrade.FFTCache — Type

struct FFTCache{A<:FFTAlg, P, I, Pu, S} <: Comrade.AbstractCache

The cache used when the FFT algorithm is used to compute visibilties. This is an internal type and is not part of the public API

Comrade.NFFTAlg — Type

NFFTAlg

Uses a non-uniform FFT to compute the visibilities. You can optionally pass uv which are the uv positions you will compute the NFFT at. This can allow for the NFFT plan to be cached improving performance

Fields

padfac: Amount to pad the image

m: Kernel size parameters. This controls the accuracy of NFFT you do not usually need to change this

σ: Over sampling factor. This controls the accuracy of NFFT you do not usually need to change this.

window: Window function for the NFFT. You do not usually need to change this

precompute: NFFT interpolation algorithm. TENSOR is the fastest but takes the longest to precompute

blocking: Flag blcok partioning should be used to speed up computation

sortNodes: Flag if the node should be sorted in a lexicographic way

storeDeconvolutionIdx: Flag if the deconvolve indices should be stored, Currently required for GPU

fftflags: Flag passed to inner AbstractFFT. The fastest FFTW is FFTW.MEASURE but takes the longest to precompute

Comrade.NUFTCache — Type

struct NUFTCache{A, P, M, PI, I} <: Comrade.AbstractCache

Internal type used to store the cache for a non-uniform Fourier transform (NUFT).

The user should instead create this using the create_cache function.

Comrade.ObservedNUFT — Type

struct ObservedNUFT{A<:Comrade.NUFT, T} <: Comrade.NUFT

Container type for a non-uniform Fourier transform (NUFT). This stores the uv-positions that the model will be sampled at in the Fourier domain, allowing certain transformtion factors (e.g., NUFT matrix) to be cached.

This is an internal type, an end user should instead create this using NFFTAlg or DFTAlg.

Modifiers

Comrade.basemodel — Function

basemodel(model::ModifiedModel)

Returns the ModifiedModel with the last transformation stripped.

Example

julia> basemodel(stretched(Disk(), 1.0, 2.0)) == Disk()
true

Comrade.unmodified — Function

unmodified(model::ModifiedModel)

Returns the un-modified model

Example

julia> m = stretched(rotated(Gaussian(), π/4), 2.0, 1.0)
julia> umodified(m) == Gaussian()
true

Comrade.renormed — Function

renormed(model, f)

Renormalizes the model m to have total flux f*flux(m). This can also be done directly by calling Base.:* i.e.,

julia> renormed(m, f) == f*M
true

Comrade.rotated — Function

rotated(model, ξ)

Rotates the model by an amount ξ in radians in the clockwise direction.

Comrade.posangle — Function

posangle(model)

Returns the rotation angle of the rotated model

Comrade.shifted — Function

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy in the x and y directions respectively.

Comrade.stretched — Function

stretched(model, α, β)

Stretches the model m according to the formula Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

Missing docstring.

Missing docstring for Comrade.AbstractModifier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Comrade.RenormalizedModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Comrade.RotatedModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Comrade.ShiftedModel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Comrade.StretchedModel. Check Documenter's build log for details.

Polarized Models

ComradeBase.mbreve — Function

mbreve(pimg, p)

Explicit m̆ function used for convenience.

ComradeBase.evpa — Method

evpa(pimg::AbstractPolarizedModel, p)

electric vector position angle or EVPA of the polarized model pimg at u and v

Comrade.PolarizedModel — Type

struct PolarizedModel{TI, TQ, TU, TV} <: ComradeBase.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

Fields

I: Stokes I model

Q: Stokes Q Model

U: Stokes U Model

V: Stokes V Model

Model Evaluation

For more docstrings on how to evaluate models see ComradeBase API.

Comrade.amplitude — Function

amplitude(d::EHTVisibilityDatum)

Get the amplitude of a visibility datum

amplitude(d::EHTVisibilityAmplitudeDatum)

Get the amplitude of a amplitude datum

amplitude(model, p)

Computes the visibility amplitude of model m at the coordinate p. The coordinate p is expected to have the properties U, V, and sometimes Ti and Fr.

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function.

Comrade.amplitudes — Function

amplitudes(m::AbstractModel, u::AbstractArray, v::AbstractArray)

Computes the visibility amplitudes of the model m at the coordinates p. The coordinates p are expected to have the properties U, V, and sometimes Ti and Fr.

Comrade.bispectra — Function

bispectra(m, p1, p2, p3)

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

Comrade.bispectrum — Function

bispectrum(d1::T, d2::T, d3::T) where {T<:EHTVisibilityDatum}

Finds the bispectrum of three visibilities. We will assume these form closed triangles, i.e. the phase of the bispectrum is a closure phase.

bispectrum(model, p1, p2, p3)

Computes the complex bispectrum of model m at the uv-triangle p1 -> p2 -> p3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function.

Comrade.closure_phase — Function

closure_phase(D1::EHTVisibilityDatum,
              D2::EHTVisibilityDatum,
              D3::EHTVisibilityDatum
              )

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

closure_phase(model, p1, p2, p3, p4)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function.

Comrade.closure_phases — Function

closure_phases(m,
               p1::AbstractArray
               p2::AbstractArray
               p3::AbstractArray
               )

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

closure_phases(m::AbstractModel, ac::ClosureConfig)

Computes the closure phases of the model m using the array configuration ac.

Notes

This is faster than the closure_phases(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

closure_phases(vis::AbstractArray, ac::ArrayConfiguration)

Compute the closure phases for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

Comrade.logclosure_amplitude — Function

logclosure_amplitude(model, p1, p2, p3, p4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v4 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function.

Comrade.logclosure_amplitudes — Function

logclosure_amplitudes(m::AbstractModel,
                      p1,
                      p2,
                      p3,
                      p4
                     )

Computes the log closure amplitudes of the model m at the quadrangles p1, p2, p3, p4.

logclosure_amplitudes(m::AbstractModel, ac::ClosureConfig)

Computes the log closure amplitudes of the model m using the array configuration ac.

Notes

This is faster than the logclosure_amplitudes(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

logclosure_amplitudes(vis::AbstractArray, ac::ArrayConfiguration)

Compute the log-closure amplitudes for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

Comrade.visibility — Function

visibility(d::EHTVisibilityDatum)

Return the complex visibility of the visibility datum

visibility(mimg, p)

Computes the complex visibility of model m at coordinates p. p corresponds to the coordinates of the model. These need to have the properties U, V and sometimes Ti for time and Fr for frequency.

Notes

If you want to compute the visibilities at a large number of positions consider using the visibilities.

Data Types

Comrade.amplitude — Method

amplitude(d::EHTVisibilityDatum)

Get the amplitude of a visibility datum

Comrade.amplitude — Method

amplitude(d::EHTVisibilityAmplitudeDatum)

Get the amplitude of a amplitude datum

Comrade.baselines — Function

baselines(CP::EHTClosurePhaseDatum)

Returns the baselines used for a single closure phase datum

baselines(CP::EHTLogClosureAmplitudeDatum)

Returns the baselines used for a single closure phase datum

baselines(scan::Scan)

Return the baselines for each datum in a scan

Comrade.arrayconfig — Function

arrayconfig(vis)

Extract the array configuration from a EHT observation.

Comrade.closure_phase — Method

closure_phase(D1::EHTVisibilityDatum,
              D2::EHTVisibilityDatum,
              D3::EHTVisibilityDatum
              )

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

Comrade.getdata — Function

getdata(obs::EHTObservation, s::Symbol)

Pass-through function that gets the array of s from the EHTObservation. For example say you want the times of all measurement then

getdata(obs, :time)

Comrade.getuv — Function

getuv

Get the u, v positions of the array.

Comrade.getuvtimefreq — Function

getuvtimefreq(ac)

Get the u, v, time, freq of the array as a tuple.

Comrade.scantable — Function

scantable(obs::EHTObservation)

Reorganizes the observation into a table of scans, where scan are defined by unique timestamps. To access the data you can use scalar indexing

Example

st = scantable(obs)
# Grab the first scan
scan1 = st[1]

# Acess the detections in the scan
scan1[1]

# grab e.g. the baselines
scan1[:baseline]

Comrade.stations — Function

stations(d::EHTObservation)

Get all the stations in a observation. The result is a vector of symbols.

stations(g::CalTable)

Return the stations in the calibration table

Comrade.uvpositions — Function

uvpositions(datum::AbstractVisibilityDatum)

Get the uvp positions of an inferometric datum.

Comrade.ArrayConfiguration — Type

abstract type ArrayConfiguration

This defined the abstract type for an array configuration. Namely, baseline times, SEFD's, bandwidth, observation frequencies, etc.

Comrade.ClosureConfig — Type

struct ClosureConfig{A, D} <: Comrade.ArrayConfiguration

Array config file for closure quantities. This stores the design matrix designmat that transforms from visibilties to closure products.

Fields

ac: Array configuration for visibilities

designmat: Closure design matrix

Comrade.AbstractInterferometryDatum — Type

abstract type AbstractInterferometryDatum{T}

An abstract type for all VLBI interfermetry data types. See Comrade.EHTVisibilityDatum for an example.

Comrade.ArrayBaselineDatum — Type

struct ArrayBaselineDatum{T, E, V}

A single datum of an ArrayConfiguration

Comrade.EHTObservation — Type

struct EHTObservation{F, T<:Comrade.AbstractInterferometryDatum{F}, S<:(StructArrays.StructArray{T<:Comrade.AbstractInterferometryDatum{F}}), A, N} <: Comrade.Observation{F}

The main data product type in Comrade this stores the data which can be a StructArray of any AbstractInterferometryDatum type.

Fields

data: StructArray of data productts

config: Array config holds ancillary information about array

mjd: modified julia date of the observation

ra: RA of the observation in J2000 (deg)

dec: DEC of the observation in J2000 (deg)

bandwidth: bandwidth of the observation (Hz)

source: Common source name

timetype: Time zone used.

Comrade.EHTArrayConfiguration — Type

struct EHTArrayConfiguration{F, T, S, D<:AbstractArray} <: Comrade.ArrayConfiguration

Stores all the non-visibility data products for an EHT array. This is useful when evaluating model visibilities.

Fields

bandwidth: Observing bandwith (Hz)

tarr: Telescope array file

scans: Scan times

data: A struct array of ArrayBaselineDatum holding time, freq, u, v, baselines.

Comrade.EHTCoherencyDatum — Type

struct EHTCoherencyDatum{S, B1, B2, M<:(StaticArraysCore.SArray{Tuple{2, 2}, Complex{S}, 2}), E<:(StaticArraysCore.SArray{Tuple{2, 2}, S, 2})} <: Comrade.AbstractInterferometryDatum{S}

A Datum for a single coherency matrix

Fields

measurement: coherency matrix, with entries in Jy

error: visibility uncertainty matrix, with entries in Jy

U: x-direction baseline length, in λ

V: y-direction baseline length, in λ

T: Timestamp, in hours

F: Frequency, in Hz

baseline: station baseline codes

polbasis: polarization basis for each station

Comrade.EHTClosurePhaseDatum — Type

struct EHTClosurePhaseDatum{S<:Number} <: Comrade.ClosureProducts{S<:Number}

A Datum for a single closure phase.

Fields

measurement: closure phase (rad)

error: error of the closure phase assuming the high-snr limit

U1: u (λ) of first station

V1: v (λ) of first station

U2: u (λ) of second station

V2: v (λ) of second station

U3: u (λ) of third station

V3: v (λ) of third station

T: Measured time of closure phase in hours

F: Measured frequency of closure phase in Hz

triangle: station baselines used

Comrade.EHTLogClosureAmplitudeDatum — Type

struct EHTLogClosureAmplitudeDatum{S<:Number} <: Comrade.ClosureProducts{S<:Number}

A Datum for a single log closure amplitude.

measurement: log-closure amplitude

error: log-closure amplitude error in the high-snr limit

U1: u (λ) of first station

V1: v (λ) of first station

U2: u (λ) of second station

V2: v (λ) of second station

U3: u (λ) of third station

V3: v (λ) of third station

U4: u (λ) of fourth station

V4: v (λ) of fourth station

T: Measured time of closure phase in hours

F: Measured frequency of closure phase in Hz

quadrangle: station codes for the quadrangle

Comrade.EHTVisibilityDatum — Type

struct EHTVisibilityDatum{S<:Number} <: Comrade.AbstractVisibilityDatum{S<:Number}

A struct holding the information for a single measured visibility.

measurement: real component of the visibility (Jy)

error: error of the visibility (Jy)

U: u position of the data point in λ

V: v position of the data point in λ

T: time of the data point in (Hr)

F: frequency of the data point (Hz)

baseline: station baseline codes

Comrade.EHTVisibilityAmplitudeDatum — Type

struct EHTVisibilityAmplitudeDatum{S<:Number} <: Comrade.AbstractVisibilityDatum{S<:Number}

A struct holding the information for a single measured visibility amplitude.

FIELDS

measurement: amplitude (Jy)

error: error of the visibility amplitude (Jy)

U: u position of the data point in λ

V: v position of the data point in λ

T: time of the data point in (Hr)

F: frequency of the data point (Hz)

baseline: station baseline codes

Comrade.Scan — Type

struct Scan{T, I, S}

Composite type that holds information for a single scan of the telescope.

Fields

time: Scan time

index: Scan indices which are (scan index, data start index, data end index)

scan: Scan data usually a StructArray of a <:AbstractVisibilityDatum

Comrade.ScanTable — Type

struct ScanTable{O<:Union{Comrade.ArrayConfiguration, Comrade.Observation}, T, S}

Wraps EHTObservation in a table that separates the observation into scans. This implements the table interface. You can access scans by directly indexing into the table. This will create a view into the table not copying the data.

Example

julia> st = scantable(obs)
julia> st[begin] # grab first scan
julia> st[end]   # grab last scan

eht-imaging interface

Comrade.extract_coherency — Function

extract_coherency(obs)

Extracts the coherency matrix from an ehtim observation object

This grabs the raw data object from the obs object. Any keyword arguments are ignored.

Returns an EHTObservation with coherency matrix

Comrade.extract_amp — Function

extract_amp(obs)

Extracts the visibility amplitudes from an ehtim observation object.

Any valid keyword arguments to add_amp in ehtim can be passed through extract_amp.

Returns an EHTObservation with visibility amplitude data

Comrade.extract_cphase — Function

extract_cphase(obs)

Extracts the closure phases from an ehtim observation object

Any valid keyword arguments to add_cphase in ehtim can be passed through extract_cphase.

Returns an EHTObservation with closure phases datums

Special Keyword arguments:

count: How the closures are formed, the available options are "min-correct", "min", "max"

cut_trivial: Cut the trivial triangles from the closures

uvmin: The flag to decide what are trivial triangles. Any baseline with ||(u,v)|| < uvmin are removed.

kwargs...: Other arguments are forwarded to eht-imaging.

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_lcamp — Function

extract_lcamp(obs)

Extracts the log-closure amp. from an ehtim observation object

Any valid keyword arguments to add_logcamp in ehtim can be passed through extract_lcamp.

Special Keyword arguments:

count: How the closures are formed, the available options are "min-correct", "min", "max"

kwargs...: Other arguments are forwarded to eht-imaging.

Returns an EHTObservation with log-closure amp. datums

Warning

The count keyword argument is treated specially in Comrade. The default option is "min-correct" and should almost always be used. This option construct a minimal set of closure phases that is valid even when the array isn't fully connected. For testing and legacy reasons we ehtim other count options are also included. However, the current ehtim count="min" option is broken and does construct proper minimal sets of closure quantities if the array isn't fully connected.

Comrade.extract_vis — Function

extract_vis(obs)

Extracts the complex visibilities from an ehtim observation object

This grabs the raw data object from the obs object. Any keyword arguments are ignored.

Returns an EHTObservation with complex visibility data

Comrade.load_ehtim_uvfits — Function

load_ehtim_uvfits(uvfile, arrayfile=nothing; kwargs...)

Load a uvfits file with eht-imaging and returns a eht-imaging Obsdata object. You can optionally pass an array file as well that will load additional information such at the telescopes field rotation information with the arrayfile. This is expected to be an eht-imaging produced array or antenna file.

Comrade.load_ehtim — Function

load_ehtim()

Loads the eht-imaging library and stores it in the exported ehtim variable.

Notes

This will fail if ehtim isn't installed in the python installation that PyCall references.

Comrade.scan_average — Function

scan_average(obs; homogenize=true)

This homogenizes the scan times for an eht-imaging Obsdata object. This is needed because eht-imaging has a bug that will sometimes create very small scans and this can mess up both the closure construction and the gain scan times. Note that this is only a problem if we are fitting scan averaged data.

Bayesian Tools

Posterior Constructions

HypercubeTransform.ascube — Function

ascube(post::Posterior)

Construct a flattened version of the posterior where the parameters are transformed to live in (0, 1), i.e. the unit hypercube.

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = ascube(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical NestedSampling methods, i.e. ComradeNested. For the transformation to unconstrained space see asflat

HypercubeTransform.asflat — Function

asflat(post::Posterior)

Construct a flattened version of the posterior where the parameters are transformed to live in (-∞, ∞).

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = ascube(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical MCMC methods, i.e. ComradeAHMC. For the transformation to the unit hypercube see ascube

ParameterHandling.flatten — Function

flatten(post::Posterior)

Construct a flattened version of the posterior but do not transform to any space, i.e. use the support specified by the prior.

This returns a TransformedPosterior that obeys the DensityInterface and can be evaluated in the usual manner, i.e. logdensityof. Note that the transformed posterior automatically includes the terms log-jacobian terms of the transformation.

Example

julia> tpost = flatten(post)
julia> x0 = prior_sample(tpost)
julia> logdensityof(tpost, x0)

Notes

This is the transform that should be used if using typical MCMC methods, i.e. ComradeAHMC. For the transformation to the unit hypercube see ascube

TransformVariables.inverse — Function

inverse(posterior::TransformedPosterior, x)

Transforms the value y from parameter space to the transformed space (e.g. unit hypercube if using ascube).

For the inverse transform see transform

Comrade.prior_sample — Function

prior_sample([rng::AbstractRandom], post::Posterior, args...)

Samples the prior distribution from the posterior. The args... are forwarded to the Base.rand method.

prior_sample([rng::AbstractRandom], post::Posterior)

Returns a single sample from the prior distribution.

Comrade.likelihood — Function

likelihood(d::ConditionedLikelihood, μ)

Returns the likelihood of the model, with parameters μ. That is, we return the distribution of the data given the model parameters μ. This is an actual probability distribution.

StatsBase.sample — Method

sample(post::Posterior, sampler::S, args...; init_params=nothing, kwargs...)

Sample a posterior post using the sampler. You can optionally pass the starting location of the sampler using init_params, otherwise a random draw from the prior will be used.

TransformVariables.transform — Function

transform(posterior::TransformedPosterior, x)

Transforms the value x from the transformed space (e.g. unit hypercube if using ascube) to parameter space which is usually encoded as a NamedTuple.

For the inverse transform see inverse

Comrade.MultiRadioLikelihood — Type

MultiRadioLikelihood(lklhd1, lklhd2, ...)

Combines multiple likelihoods into one object that is useful for fitting multiple days/frequencies.

julia> lklhd1 = RadioLikelihood(dcphase1, dlcamp1)
julia> lklhd2 = RadioLikelihood(dcphase2, dlcamp2)
julia> MultiRadioLikelihood(lklhd1, lklhd2)

Comrade.Posterior — Type

Posterior(lklhd, prior)

Creates a Posterior density that follows obeys DensityInterface. The lklhd object is expected to be a VLB object. For instance, these can be created using RadioLikelihood. prior is expected to be a NamedTuple of distributions that reflect the priors on the parameters you are considering.

Notes

Since this function obeys DensityInterface you can evaluate it with

julia> ℓ = logdensityof(post)
julia> ℓ(x)

or using the 2-argument version directly

julia> logdensityof(post, x)

where post::Posterior.

To generate random draws from the prior see the prior_sample function.

Comrade.TransformedPosterior — Type

struct TransformedPosterior{P<:Posterior, T} <: Comrade.AbstractPosterior

A transformed version of a Posterior object. This is an internal type that an end user shouldn't have to directly construct. To construct a transformed posterior see the asflat, ascube, and flatten docstrings.

Comrade.RadioLikelihood — Type

RadioLikelihood(model, data1, data2, ...)

Forms a radio likelihood from a set of data products. These data products must share the same array data/configuration. If you want to form a likelihood from multiple arrays such as when fitting different wavelengths or days, you can combine them using MultiRadioLikelihood

Example

julia> RadioLikelihood(model, dcphase1, dlcamp1)

Comrade.IsFlat — Type

struct IsFlat

Specifies that the sampling algorithm usually expects a uncontrained transform

Comrade.IsCube — Type

struct IsCube

Specifies that the sampling algorithm usually expects a hypercube transform

Misc

Comrade.μas2rad — Function

μas2rad(x)

Converts a number from micro-arcseconds (μas) to rad

Comrade.rad2μas — Function

rad2μas(x)

Converts a number from radians to micro-arcseconds (μas)

Comrade.load — Function

Comrade.load(fitsfile::String, IntensityMap)

This loads in a fits file that is more robust to the various imaging algorithms in the EHT, i.e. is works with clean, smili, eht-imaging. The function returns an tuple with an intensitymap and a second named tuple with ancillary information about the image, like the source name, location, mjd, and radio frequency.

Comrade.save — Function

Comrade.save(file::String, img::IntensityMap, obs)

Saves an image to a fits file. You can optionally pass an EHTObservation so that ancillary information will be added.

Comrade.NonAnalyticTest — Type

`NonAnalyticTest`

An internal model used primarly for testing. Any model passed to it will be interpreted as not having an analytic Fourier transform.

Internal (Not Public API)

Comrade.scale_uv — Function

scale_image(model::AbstractModifier, u, u)

Returns a number on how to scale the image visibility at u v for an modified model

Comrade.scale_image — Function

scale_image(model::AbstractModifier, x, y)

Returns a number of how to to scale the image intensity at x y for an modified model

Comrade.transform_uv — Function

transform_uv(model::AbstractModifier, u, u)

Returns a transformed u and v according to the model modifier

Comrade.transform_image — Function

transform_image(model::AbstractModifier, x, y)

Returns a transformed x and y according to the model modifier

Comrade.ThreadedModel — Type

ThreadedModel

Experimental model wrapper than enables multi-threading when evaluating intensitymap

Comrade.extract_FRs — Function

extract_FRs

Extracts the feed rotation Jones matrices (returned as a JonesPair) from an EHT observation obs.

Warning

eht-imaging can sometimes pre-rotate the coherency matrices. As a result the field rotation can sometimes be applied twice. To compensate for this we have added a ehtim_fr_convention which will fix this.

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020