ComradeBase API

Contents

ComradeBase API

Contents

Index

Model API

Image Pulses

Polarization

Index

ComradeBase.AbstractModel

ComradeBase.BSplinePulse

ComradeBase.CoherencyMatrix

ComradeBase.DeltaPulse

ComradeBase.DensityAnalytic

ComradeBase.IntensityMap

ComradeBase.IsAnalytic

ComradeBase.IsPrimitive

ComradeBase.NotAnalytic

ComradeBase.NotPrimitive

ComradeBase.PrimitiveTrait

ComradeBase.Pulse

ComradeBase.SingleStokes

ComradeBase.SqExpPulse

ComradeBase.StokesVector

ComradeBase.evpa

ComradeBase.flux

ComradeBase.imanalytic

ComradeBase.intensity_point

ComradeBase.isprimitive

ComradeBase.m̆

ComradeBase.radialextent

ComradeBase.visanalytic

ComradeBase.visibility_point

Model API

ComradeBase.AbstractModel — Type

AbstractModel

The Comrade abstract model type. To instantiate your own model type you should subtybe from this model. Additionally you need to implement the following methods to satify the interface:

Mandatory Methods

isprimitive: defines whether a model is standalone or is defined in terms of other models. is the model is primitive then this should return IsPrimitive() otherwise it returns NotPrimitive()

visanalytic: defines whether the model visibilities can be computed analytically. If yes then this should return IsAnalytic() and the user must to define visibility_point. If not analytic then visanalytic should return NotAnalytic().

imanalytic: defines whether the model intensities can be computed pointwise. If yes

then this should return IsAnalytic() and the user must to define intensity_point. If not analytic then imanalytic should return NotAnalytic().

radialextent: Provides a estimate of the radial extent of the model in the image domain. This is used for estimating the size of the image, and for plotting.

flux: Returns the total flux of the model.

Optional Methods:

intensity_point: Defines how to compute model intensities pointwise. Note this is must be defined if imanalytic(::Type{YourModel})==IsAnalytic().

visibility_point: Defines how to compute model visibilties pointwise. Note this is must be defined if visanalytic(::Type{YourModel})==IsAnalytic().

_visibilities: Vectorized version of visibility_point if you can gain additional speed

intensitymap: Computes the whole image of the model

intensitymap!: Inplace version of intensitymap

ComradeBase.flux — Function

flux(im::AbstractIntensityMap)

Computes the flux of a intensity map

flux(im)

ComradeBase.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive Comrade model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of Comrade they need to specify if it is primitive

struct MyPrimitiveModel end
ComradeBase.isprimitive(::Type{MyModel}) = ComradeBase.IsPrimitive()

ComradeBase.visanalytic — Function

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ComradeBase.imanalytic — Function

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ComradeBase.radialextent — Function

radialextent(model::AbstractModel)

Provides an estimate of the radial size/extent of the model. This is used internally to estimate image size when plotting and using modelimage

ComradeBase.visibility_point — Function

visibility_point(model::AbstractModel, u, v, args...)

Function that computes the pointwise visibility. This must be implemented in the model interface if visanalytic(::Type{MyModel}) == IsAnalytic()

ComradeBase.intensity_point — Function

intensity_point(model::AbstractModel, x, y, args...)

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ComradeBase.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into Comrade more easily.

ComradeBase.IsPrimitive — Type

struct IsPrimitive

Trait for primitive model

ComradeBase.NotPrimitive — Type

struct NotPrimitive

Trait for not-primitive model

ComradeBase.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ComradeBase.IsAnalytic — Type

struct IsAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ComradeBase.NotAnalytic — Type

struct NotAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ComradeBase.IntensityMap — Type

struct IntensityMap{T, S<:(AbstractMatrix), F, K<:ComradeBase.Pulse} <: ComradeBase.AbstractIntensityMap{T, S<:(AbstractMatrix)}

Image array type. This is an Matrix with a number of internal fields to describe the field of view, pixel size, and the pulse function that makes the image a continuous quantity.

To use it you just specify the array and the field of view/pulse $julia img = IntensityMap(zeros(512,512), 100.0, 100.0, DeltaPulse)$`

Fields

im

Image matrix

fovx

field of view in x direction

fovy

field of view in y direction

psizex

pixel size in the x direction

psizey

pixel size in the y direction

pulse

pulse function that turns the image grid into a continuous object

Image Pulses

ComradeBase.Pulse — Type

Pulse Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

ComradeBase.DeltaPulse — Type

struct DeltaPulse{T} <: ComradeBase.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

ComradeBase.SqExpPulse — Type

struct SqExpPulse{T} <: ComradeBase.Pulse

Normalized square exponential kernel, i.e. a Gaussian. Note the smoothness is modfied with ϵ which is the inverse variance in units of 1/pixels².

ComradeBase.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

Polarization

ComradeBase.StokesVector — Type

struct StokesVector{T} <: StaticArrays.FieldVector{4, T}

Static vector that holds the stokes parameters of a polarized complex visibility

To convert between a StokesVector and CoherencyMatrix use the convert function

convert(::CoherencyMatrix, StokesVector(1.0, 0.1, 0.1, 0.4))

ComradeBase.CoherencyMatrix — Type

struct CoherencyMatrix{T} <: StaticArrays.FieldMatrix{2, 2, T}

Static matrix that holds construct the coherency matrix of a polarized complex visibility

To convert between a StokesVector and CoherencyMatrix use the convert function

convert(::StokesVector, CoherencyMatrix(1.0, 0.1, 0.1, 0.4))

ComradeBase.evpa — Function

evpa(m)

Compute the evpa of a stokes vector or cohereny matrix.

evpa(pimg::AbstractPolarizedModel, u, v)

electric vector position angle or EVPA of the polarized model pimg at u and v

ComradeBase.m̆ — Function

m̆(m)

Compute the fractional linear polarization of a stokes vector or coherency matrix

m̆(pimg::AbstractPolarizedModel, u, v)

Computes the fractional linear polarization in the visibility domain

m̆ = (Q + iU)/I

To create the symbol type m\breve in the REPL or use the mbreve function.

ComradeBase.SingleStokes — Type

struct SingleStokes{M, S} <: ComradeBase.AbstractModel

Helper function that converts a model from something that compute polarized images to just a single stokes parameter. This is useful if you just want to fit a single stokes parameter.