ComradeBase API

Contents

ComradeBase API

Contents

Index

Model API

Polarization

Index

ComradeBase.AbstractModel

ComradeBase.AbstractPolarizedModel

ComradeBase.DensityAnalytic

ComradeBase.GriddedKeys

ComradeBase.IntensityMap

ComradeBase.IntensityMap

ComradeBase.IsAnalytic

ComradeBase.IsPrimitive

ComradeBase.MinimalHeader

ComradeBase.NoHeader

ComradeBase.NotAnalytic

ComradeBase.NotPrimitive

ComradeBase.PrimitiveTrait

ComradeBase.StokesIntensityMap

PolarizedTypes.CirBasis

PolarizedTypes.CoherencyMatrix

PolarizedTypes.ElectricFieldBasis

PolarizedTypes.LPol

PolarizedTypes.LinBasis

PolarizedTypes.PolBasis

PolarizedTypes.RPol

PolarizedTypes.StokesParams

PolarizedTypes.XPol

PolarizedTypes.YPol

ComradeBase.amplitude

ComradeBase.amplitudes

ComradeBase.axisdims

ComradeBase.bispectra

ComradeBase.bispectrum

ComradeBase.centroid

ComradeBase.closure_phase

ComradeBase.closure_phases

ComradeBase.dims

ComradeBase.fieldofview

ComradeBase.flux

ComradeBase.header

ComradeBase.imagegrid

ComradeBase.imagepixels

ComradeBase.imanalytic

ComradeBase.intensity_point

ComradeBase.intensitymap

ComradeBase.intensitymap!

ComradeBase.intensitymap_analytic

ComradeBase.intensitymap_analytic!

ComradeBase.intensitymap_numeric

ComradeBase.intensitymap_numeric!

ComradeBase.ispolarized

ComradeBase.isprimitive

ComradeBase.load

ComradeBase.logclosure_amplitude

ComradeBase.logclosure_amplitudes

ComradeBase.named_dims

ComradeBase.phasecenter

ComradeBase.pixelsizes

ComradeBase.radialextent

ComradeBase.save

ComradeBase.second_moment

ComradeBase.stokes

ComradeBase.visanalytic

ComradeBase.visibilities

ComradeBase.visibilities!

ComradeBase.visibilities_analytic

ComradeBase.visibilities_analytic!

ComradeBase.visibilities_numeric

ComradeBase.visibilities_numeric!

ComradeBase.visibility

ComradeBase.visibility_point

PolarizedTypes.basis_components

PolarizedTypes.basis_transform

PolarizedTypes.innerprod

Model API

ComradeBase.flux — Function

flux(im::IntensityMap)
flux(img::StokesIntensityMap)

Computes the flux of a intensity map

ComradeBase.visibility — Function

visibility(mimg, p)

Computes the complex visibility of model m at coordinates p. p corresponds to the coordinates of the model. These need to have the properties U, V and sometimes Ti for time and Fr for frequency.

Notes

If you want to compute the visibilities at a large number of positions consider using the visibilities.

visibility(d::EHTVisibilityDatum)

Return the complex visibility of the visibility datum

ComradeBase.visibilities — Function

visibilities(model::AbstractModel, args...)

Computes the complex visibilities at the locations given by args...

ComradeBase.visibilities! — Function

visibilities!(vis::AbstractArray, model::AbstractModel, args...)

Computes the complex visibilities vis in place at the locations given by args...

ComradeBase.intensitymap — Function

intensitymap(model::AbstractModel, p::AbstractDims)

Computes the intensity map of model. For the inplace version see intensitymap!

ComradeBase.intensitymap! — Function

intensitymap!(buffer::AbstractDimArray, model::AbstractModel)

Computes the intensity map of model by modifying the buffer

ComradeBase.IntensityMap — Type

IntensityMap(data::AbstractArray, dims::NamedTuple)
IntensityMap(data::AbstractArray, grid::AbstractDims)

Constructs an intensitymap using the image dimensions given by dims. This returns a KeyedArray with keys given by an ImageDimensions object.

dims = (X=range(-10.0, 10.0, length=100), Y = range(-10.0, 10.0, length=100),
        T = [0.1, 0.2, 0.5, 0.9, 1.0], F = [230e9, 345e9]
        )
imgk = IntensityMap(rand(100,100,5,1), dims)

ComradeBase.amplitude — Method

amplitude(model, p)

Computes the visibility amplitude of model m at the coordinate p. The coordinate p is expected to have the properties U, V, and sometimes Ti and Fr.

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function.

ComradeBase.amplitudes — Function

amplitudes(m::AbstractModel, u::AbstractArray, v::AbstractArray)

Computes the visibility amplitudes of the model m at the coordinates p. The coordinates p are expected to have the properties U, V, and sometimes Ti and Fr.

ComradeBase.bispectrum — Function

bispectrum(model, p1, p2, p3)

Computes the complex bispectrum of model m at the uv-triangle p1 -> p2 -> p3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function.

bispectrum(d1::T, d2::T, d3::T) where {T<:EHTVisibilityDatum}

Finds the bispectrum of three visibilities. We will assume these form closed triangles, i.e. the phase of the bispectrum is a closure phase.

ComradeBase.bispectra — Function

bispectra(m, p1, p2, p3)

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

ComradeBase.closure_phase — Function

closure_phase(model, p1, p2, p3, p4)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function.

closure_phase(D1::EHTVisibilityDatum,
              D2::EHTVisibilityDatum,
              D3::EHTVisibilityDatum
              )

Computes the closure phase of the three visibility datums.

Notes

We currently use the high SNR Gaussian error approximation for the closure phase. In the future we may use the moment matching from Monte Carlo sampling.

ComradeBase.closure_phases — Function

closure_phases(m,
               p1::AbstractArray
               p2::AbstractArray
               p3::AbstractArray
               )

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

closure_phases(m::AbstractModel, ac::ClosureConfig)

Computes the closure phases of the model m using the array configuration ac.

Notes

This is faster than the closure_phases(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

closure_phases(vis::AbstractArray, ac::ArrayConfiguration)

Compute the closure phases for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

ComradeBase.logclosure_amplitude — Function

logclosure_amplitude(model, p1, p2, p3, p4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v4 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function.

ComradeBase.logclosure_amplitudes — Function

logclosure_amplitudes(m::AbstractModel,
                      p1,
                      p2,
                      p3,
                      p4
                     )

Computes the log closure amplitudes of the model m at the quadrangles p1, p2, p3, p4.

logclosure_amplitudes(m::AbstractModel, ac::ClosureConfig)

Computes the log closure amplitudes of the model m using the array configuration ac.

Notes

This is faster than the logclosure_amplitudes(m, u1, v1, ...) method since it only computes as many visibilities as required thanks to the closure design matrix formalism from Blackburn et al.[1]

logclosure_amplitudes(vis::AbstractArray, ac::ArrayConfiguration)

Compute the log-closure amplitudes for a set of visibilities and an array configuration

Notes

This uses a closure design matrix for the computation.

Model Interface

ComradeBase.AbstractModel — Type

AbstractModel

The Comrade abstract model type. To instantiate your own model type you should subtybe from this model. Additionally you need to implement the following methods to satify the interface:

Mandatory Methods

isprimitive: defines whether a model is standalone or is defined in terms of other models. is the model is primitive then this should return IsPrimitive() otherwise it returns NotPrimitive()

visanalytic: defines whether the model visibilities can be computed analytically. If yes then this should return IsAnalytic() and the user must to define visibility_point. If not analytic then visanalytic should return NotAnalytic().

imanalytic: defines whether the model intensities can be computed pointwise. If yes then this should return IsAnalytic() and the user must to define intensity_point. If not analytic then imanalytic should return NotAnalytic().

radialextent: Provides a estimate of the radial extent of the model in the image domain. This is used for estimating the size of the image, and for plotting.

flux: Returns the total flux of the model.

intensity_point: Defines how to compute model intensities pointwise. Note this is must be defined if imanalytic(::Type{YourModel})==IsAnalytic().

visibility_point: Defines how to compute model visibilties pointwise. Note this is must be defined if visanalytic(::Type{YourModel})==IsAnalytic().

Optional Methods:

ispolarized: Specified whether a model is intrinsically polarized (returns IsPolarized()) or is not (returns NotPolarized()), by default a model is NotPolarized()

visibilities_analytic: Vectorized version of visibility_point for models where visanalytic returns IsAnalytic()

visibilities_numeric: Vectorized version of visibility_point for models where visanalytic returns NotAnalytic() typically these are numerical FT's

intensitymap_analytic: Computes the entire image for models where imanalytic returns IsAnalytic()

intensitymap_numeric: Computes the entire image for models where imanalytic returns NotAnalytic()

intensitymap_analytic!: Inplace version of intensitymap

intensitymap_numeric!: Inplace version of intensitymap

ComradeBase.isprimitive — Function

isprimitive(::Type)

Dispatch function that specifies whether a type is a primitive Comrade model. This function is used for dispatch purposes when composing models.

Notes

If a user is specifying their own model primitive model outside of Comrade they need to specify if it is primitive

struct MyPrimitiveModel end
ComradeBase.isprimitive(::Type{MyModel}) = ComradeBase.IsPrimitive()

ComradeBase.visanalytic — Function

visanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in Fourier domain, i.e. we can evaluate its fourier transform at an arbritrary point.

If IsAnalytic() then it will try to call visibility_point to calculate the complex visibilities. Otherwise it fallback to using the FFT that works for all models that can compute an image.

ComradeBase.imanalytic — Function

imanalytic(::Type{<:AbstractModel})

Determines whether the model is pointwise analytic in the image domain, i.e. we can evaluate its intensity at an arbritrary point.

If IsAnalytic() then it will try to call intensity_point to calculate the intensity.

ComradeBase.ispolarized — Function

ispolarized(::Type)

Trait function that defines whether a model is polarized or not.

ComradeBase.radialextent — Function

radialextent(model::AbstractModel)

Provides an estimate of the radial size/extent of the model. This is used internally to estimate image size when plotting and using modelimage

ComradeBase.PrimitiveTrait — Type

abstract type PrimitiveTrait

This trait specifies whether the model is a primitive

Notes

This will likely turn into a trait in the future so people can inject their models into Comrade more easily.

ComradeBase.IsPrimitive — Type

struct IsPrimitive

Trait for primitive model

ComradeBase.NotPrimitive — Type

struct NotPrimitive

Trait for not-primitive model

ComradeBase.DensityAnalytic — Type

DensityAnalytic

Internal type for specifying the nature of the model functions. Whether they can be easily evaluated pointwise analytic. This is an internal type that may change.

ComradeBase.IsAnalytic — Type

struct IsAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has a analytic fourier transform and/or image.

ComradeBase.NotAnalytic — Type

struct NotAnalytic <: ComradeBase.DensityAnalytic

Defines a trait that a states that a model is analytic. This is usually used with an abstract model where we use it to specify whether a model has does not have a easy analytic fourier transform and/or intensity function.

ComradeBase.visibility_point — Function

visibility_point(model::AbstractModel, p)

Function that computes the pointwise visibility. This must be implemented in the model interface if visanalytic(::Type{MyModel}) == IsAnalytic()

ComradeBase.visibilities_analytic — Function

visibilties_analytic(model, u, v, time, freq)

Computes the visibilties of a model using using the analytic visibility expression given by visibility_point.

ComradeBase.visibilities_analytic! — Function

visibilties_analytic!(vis, model, u, v, time, freq)

Computes the visibilties of a model in-place, using using the analytic visibility expression given by visibility_point.

ComradeBase.visibilities_numeric — Function

visibilties_numeric(model, u, v, time, freq)

Computes the visibilties of a model using a numerical fourier transform. Note that none of these are implemented in ComradeBase. For implementations please see Comrade.

ComradeBase.visibilities_numeric! — Function

visibilties_numeric!(vis, model, u, v, time, freq)

Computes the visibilties of a model in-place using a numerical fourier transform. Note that none of these are implemented in ComradeBase. For implementations please see Comrade.

ComradeBase.intensity_point — Function

intensity_point(model::AbstractModel, p)

Function that computes the pointwise intensity if the model has the trait in the image domain IsAnalytic(). Otherwise it will use construct the image in visibility space and invert it.

ComradeBase.intensitymap_analytic — Function

intensitymap_analytic(m::AbstractModel, p::AbstractDims)

Computes the IntensityMap of a model m using the image dimensions p by broadcasting over the analytic intensity_point method.

ComradeBase.intensitymap_analytic! — Function

intensitymap_analytic!(img::IntensityMap, m::AbstractModel)
intensitymap_analytic!(img::StokesIntensityMap, m::AbstractModel)

Updates the img using the model m by broadcasting over the analytic intensity_point method.

ComradeBase.intensitymap_numeric — Function

intensitymap_numeric(m::AbstractModel, p::AbstractDims)

Computes the IntensityMap of a model m at the image positions p using a numerical method. This has to be specified uniquely for every model m if imanalytic(typeof(m)) === NotAnalytic(). See Comrade.jl for example implementations.

ComradeBase.intensitymap_numeric! — Function

intensitymap_numeric!(img::IntensityMap, m::AbstractModel)
intensitymap_numeric!(img::StokesIntensityMap, m::AbstractModel)

Updates the img using the model m using a numerical method. This has to be specified uniquely for every model m if imanalytic(typeof(m)) === NotAnalytic(). See Comrade.jl for example implementations.

Image Types

ComradeBase.IntensityMap — Method

IntensityMap(data::AbstractArray, dims::NamedTuple)
IntensityMap(data::AbstractArray, grid::AbstractDims)

Constructs an intensitymap using the image dimensions given by dims. This returns a KeyedArray with keys given by an ImageDimensions object.

dims = (X=range(-10.0, 10.0, length=100), Y = range(-10.0, 10.0, length=100),
        T = [0.1, 0.2, 0.5, 0.9, 1.0], F = [230e9, 345e9]
        )
imgk = IntensityMap(rand(100,100,5,1), dims)

ComradeBase.StokesIntensityMap — Type

struct StokesIntensityMap{T, N, SI, SQ, SU, SV}

General struct that holds intensity maps for each stokes parameter. Each image I, Q, U, V must share the same axis dimensions. This type also obeys much of the usual array interface in Julia. The following methods have been implemented:

size

eltype (returns StokesParams)

ndims

getindex

setindex!

pixelsizes

fieldofview

imagepixels

imagegrid

stokes

Warning

This may eventually be phased out for IntensityMaps whose base types are StokesParams, but currently we use this for speed reasons with Zygote.

ComradeBase.imagepixels — Function

imagepixels(img::IntensityMap)
imagepixels(img::IntensityMapTypes)

Returns a abstract spatial dimension with the image pixels locations X and Y.

ComradeBase.GriddedKeys — Type

struct GriddedKeys{N, G, Hd, T} <: ComradeBase.AbstractDims{N, T}

This struct holds the dimensions that the EHT expect. The first type parameter N defines the names of each dimension. These names are usually one of - (:X, :Y, :T, :F) - (:X, :Y, :F, :T) - (:X, :Y) # spatial only where :X,:Y are the RA and DEC spatial dimensions respectively, :T is the the time direction and :F is the frequency direction.

Fieldnames

dims

header

Notes

Warning it is rare you need to access this constructor directly. Instead use the direct IntensityMap function.

ComradeBase.dims — Function

dims(g::AbstractDims)

Returns a tuple containing the dimensions of g. For a named version see ComradeBase.named_dims

ComradeBase.named_dims — Function

named_dims(g::AbstractDims)

Returns a named tuple containing the dimensions of g. For a unnamed version see dims

ComradeBase.axisdims — Function

axisdims(img::IntensityMap)

Returns the keys of the IntensityMap as the actual internal AbstractDims object.

ComradeBase.stokes — Function

stokes(m::AbstractPolarizedModel, p::Symbol)

Extract the specific stokes component p from the polarized model m

ComradeBase.imagegrid — Function

imagegrid(k::IntensityMap)

Returns the grid the IntensityMap is defined as. Note that this is unallocating since it lazily computes the grid. The grid is an example of a KeyedArray and works similarly. This is useful for broadcasting a model across an abritrary grid.

ComradeBase.fieldofview — Function

fieldofview(img::IntensityMap)
fieldofview(img::IntensityMapTypes)

Returns a named tuple with the field of view of the image.

ComradeBase.pixelsizes — Function

pixelsizes(img::IntensityMap)
pixelsizes(img::IntensityMapTypes)

Returns a named tuple with the spatial pixel sizes of the image.

ComradeBase.phasecenter — Function

phasecenter(img::IntensityMap)
phasecenter(img::StokesIntensitymap)

Computes the phase center of an intensity map. Note this is the pixels that is in the middle of the image.

ComradeBase.centroid — Function

centroid(im::AbstractIntensityMap)

Computes the image centroid aka the center of light of the image.

For polarized maps we return the centroid for Stokes I only.

ComradeBase.second_moment — Function

second_moment(im::AbstractIntensityMap; center=true)

Computes the image second moment tensor of the image. By default we really return the second cumulant or centered second moment, which is specified by the center argument.

For polarized maps we return the second moment for Stokes I only.

second_moment(im::AbstractIntensityMap; center=true)

Computes the image second moment tensor of the image. By default we really return the second cumulant or centered second moment, which is specified by the center argument.

ComradeBase.header — Function

header(g::AbstractDims)

Returns the headerinformation of the dimensions g

header(img::IntensityMap)

Retrieves the header of an IntensityMap

ComradeBase.NoHeader — Type

NoHeader

ComradeBase.MinimalHeader — Type

MinimalHeader{T}

A minimal header type for ancillary image information.

Fields

source: Common source name

ra: Right ascension of the image in degrees (J2000)

dec: Declination of the image in degrees (J2000)

mjd: Modified Julian Date in days

frequency: Frequency of the image in Hz

ComradeBase.load — Function

ComradeBase.load(fitsfile::String, IntensityMap)

This loads in a fits file that is more robust to the various imaging algorithms in the EHT, i.e. is works with clean, smili, eht-imaging. The function returns an tuple with an intensitymap and a second named tuple with ancillary information about the image, like the source name, location, mjd, and radio frequency.

ComradeBase.save — Function

ComradeBase.save(file::String, img::IntensityMap, obs)

Saves an image to a fits file. You can optionally pass an EHTObservation so that ancillary information will be added.

Polarization

ComradeBase.AbstractPolarizedModel — Type

abstract type AbstractPolarizedModel <: ComradeBase.AbstractModel

Type the classifies a model as being intrinsically polarized. This means that any call to visibility must return a StokesParams to denote the full stokes polarization of the model.

PolarizedTypes.StokesParams — Type

struct StokesParams{T} <: StaticArraysCore.FieldVector{4, T}

Static vector that holds the stokes parameters of a polarized complex visibility

To convert between a StokesParams and CoherencyMatrix use the convert function

convert(::CoherencyMatrix, StokesVector(1.0, 0.1, 0.1, 0.4))

PolarizedTypes.ElectricFieldBasis — Type

abstract type ElectricFieldBasis

An abstract type whose subtypes denote a specific electric field basis.

PolarizedTypes.RPol — Type

struct RPol <: PolarizedTypes.ElectricFieldBasis

The right circular electric field basis, i.e. a right-handed circular feed.

PolarizedTypes.LPol — Type

struct LPol <: PolarizedTypes.ElectricFieldBasis

The left circular electric field basis, i.e. a left-handed circular feed.

PolarizedTypes.XPol — Type

struct XPol <: PolarizedTypes.ElectricFieldBasis

The horizontal or X electric feed basis, i.e. the horizontal linear feed.

PolarizedTypes.YPol — Type

struct YPol <: PolarizedTypes.ElectricFieldBasis

The vertical or Y electric feed basis, i.e. the vertical linear feed.

PolarizedTypes.PolBasis — Type

struct PolBasis{B1<:Union{Missing, PolarizedTypes.ElectricFieldBasis}, B2<:Union{Missing, PolarizedTypes.ElectricFieldBasis}}

Denotes a general polarization basis, with basis vectors (B1,B2) which are typically <: Union{ElectricFieldBasis, Missing}

PolarizedTypes.CirBasis — Type

CirBasis <: PolBasis

Measurement uses the circular polarization basis, which is typically used for circular feed interferometers.

PolarizedTypes.LinBasis — Type

LinBasis <: PolBasis

Measurement uses the linear polarization basis, which is typically used for linear feed interferometers.

PolarizedTypes.CoherencyMatrix — Type

struct CoherencyMatrix{B1, B2, T} <: StaticArraysCore.FieldMatrix{2, 2, T}

Coherency matrix for a single baseline with bases B1 and B2. The two bases correspond to the type of feeds used for each telescope and should be subtypes of PolBasis. To see which bases are implemented type subtypes(Rimes.PolBasis) in the REPL.

For a circular basis the layout of the coherency matrix is

RR* RL*
LR* RR*

which can be constructed using

c = CoherencyMatrix(RR, LR, RL, LL, CirBasis())

For a linear basis the layout of the coherency matrix is

XX* XY*
YX* YY*

which can be constructed using

c = CoherencyMatrix(XX, YX, XY, YY, CirBasis())

For a mixed (e.g., circular and linear basis) the layout of the coherency matrix is

RX* RY*
LX* LY*

or e.g., linear and circular the layout of the coherency matrix is

XR* XL*
YR* YL*

These coherency matrices can be constructed using:

# Circular and linear feeds i.e., |R><X|
c = CoherencyMatrix(RX, LX, RY, LY, LinBasis(), CirBasis())
# Linear and circular feeds i.e., |X><R|
c = CoherencyMatrix(XR, YR, XL, YL, LinBasis(), CirBasis())

Missing docstring.

Missing docstring for PolarizedTypes.evpa. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PolarizedTypes.m̆. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PolarizedTypes.linearpol. Check Documenter's build log for details.

PolarizedTypes.innerprod — Function

innerprod(::Type{T}, XPol(), YPol())

Computes the complex inner product of two elements of a complex Hilbert space X and Y where base element of the output is T.

PolarizedTypes.basis_components — Function

basis_components([T=Float64,], e::ElectricFieldBasis, b::PolBasis)

Returns a static vector that contains the components of the electric field basis vector e in terms of the polarization basis b. The first argument is optionally the eltype of the static vector.

Examples

julia> basis_components(Float64, R(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 0.7071067811865475 + 0.0im
                0.0 - 0.7071067811865475im

julia> basis_components(R(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 0.7071067811865475 + 0.0im
                0.0 - 0.7071067811865475im

julia> basis_components(Float64, X(), PolBasis{XPol,Y}())
2-element StaticArraysCore.SVector{2, ComplexF64} with indices SOneTo(2):
 1.0 + 0.0im
 0.0 + 0.0im

PolarizedTypes.basis_transform — Function

basis_transform([T=Float64,], b1::PolBasis, b2::PolBasis)
basis_transform([T=Float64,], b1::PolBasis=>b2::PolBasis)

Produces the transformation matrix that transforms the vector components from basis b1 to basis b2. This means that if for example E is the circular basis then basis_transform(CirBasis=>LinBasis)E is in the linear basis. In other words the columns of the transformation matrix are the coordinate vectors of the new basis vectors in the old basis.

Example

julia> basis_transform(CirBasis()=>LinBasis())
2×2 StaticArraysCore.SMatrix{2, 2, ComplexF64, 4} with indices SOneTo(2)×SOneTo(2):
 0.707107-0.0im       0.707107-0.0im
      0.0-0.707107im       0.0+0.707107im

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020

1Blackburn L., et al "Closure Statistics in Interferometric Data" ApJ 2020