ComradeAHMC

The first choice when sampling from the model/image posterior, is AdvancedHMC ), which uses Hamiltonian Monte Carlo to sample from the posterior. Specifically, we usually use the NUTS algorithm.

The interface to AdvancedHMC is very powerful and general. To simplify the procedure for Comrade users, we have provided a thin interface. A user needs to specify a sampler and then call the sample function.

For AdvancedHMC, the user can create the sampler by calling the AHMC function. This only has one mandatory argument, the metric the sampler uses. There are currently two options:

- `DiagEuclideanMetric` which uses a diagonal metric for covariance adaptation
- `DenseEuclideanMetric` which uses a dense or full rank metric for covariance adaptation

We recommend that a user starts with DiagEuclideanMetric since the dense metric typically requires many more samples to tune correctly. The other options for AHMC (sans autodiff) specify which version of HMC to use. Our default options match the choices made by the Stan programming language. The final option to consider is the autodiff optional argument. This specifies which auto differentiation package to use. For geometric modeling, we recommend the AD.ForwardDiffBackend(), while for Bayesian Imaging, AD.ZygoteBackend().

Example

using Comrade
using ComradeAHMC

# Some stuff to create a posterior object
post # of type Comrade.Posterior

metric = DiagEuclideanMetric(dimension(post))
smplr = AHMC(metric=metric, autodiff=AD.ForwardDiffBackend())

chain, stats = sample(post, smplr, 2_000; nadapts=1_000)

API

ComradeAHMC.AHMC — Type

AdvancedHMC

Creates a sampler that uses the AdvancedHMC framework to construct an Hamiltonian Monte Carlo NUTS sampler.

The user must specify the metric they want to use. Typically we recommend DiagEuclideanMetric as a reasonable starting place. The other options are chosen to match the Stan languages defaults and should provide a good starting point. Please see the AdvancedHMC docs for more information.

Notes

For autodiff the user can either specify an AbstractDifferentiation backend. For smaller geometric models we recommend AD.ForwardDiffBackend while for imaging we recommend AD.Zygote.

Fields

struct AHMC{S,I,P,T,A,D}
    `AdvancedHMC` metric to use
    metric::S
    `AdvancedHMC` integrator
    Defaults to `AdvancedHMC.Leapfrog`
    integrator::I = Leapfrog
    HMC trajectory sampler
    Defaults to `AdvancedHMC.MultinomialTS`
    trajectory::P = MultinomialTS
    HMC termination condition
    Defaults to `AdvancedHMC.StrictGeneralisedNoUTurn`
    termination::T = StrictGeneralisedNoUTurn(10, 1000.0)
    Adaptation strategy for mass matrix and stepsize
    Defaults to `AdvancedHMC.StanHMCAdaptor`
    adaptor::A = StanHMCAdaptor
    Target acceptance rate for all trajectories on the tree
    Defaults to 0.85
    targetacc::Float64 = 0.8
    `AbstractDifferentiation` autodiff backend
    Defaults to `AbstractDifferentiation.ForwardDiffBackend()`
    autodiff::D = AD.ForwardDiffBackend()
end

StatsBase.sample — Method

AbstractMCMC.sample(post::Comrade.Posterior,
                    sampler::AHMC,
                    nsamples;
                    init_params=nothing,
                    kwargs...)

Samples the posterior post using the AdvancedHMC sampler specified by AHMC. This will run the sampler for nsamples.

To initialize the chain the user can set init_params to Vector{NamedTuple} whose elements are the starting locations for each of the nchains. If no starting location is specified nchains random samples from the prior will be chosen for the starting locations.

For possible kwargs please see the AdvancedHMC.jl docs

This returns a tuple where the first element is a TypedTable of the MCMC samples in parameter space and the second argument is a set of ancilliary information about the sampler.

Notes

This will automatically transform the posterior to the flattened unconstrained space.

StatsBase.sample — Method

AbstractMCMC.sample(post::Comrade.Posterior,
                    sampler::AHMC,
                    parallel::AbstractMCMC.AbstractMCMCEnsemble,
                    nsamples,
                    nchainsl;
                    init_params=nothing,
                    kwargs...)

Samples the posterior post using the AdvancedHMC sampler specified by AHMC. This will sample nchains copies of the posterior using the parallel scheme. Each chain will be sampled for nsamples.

To initialize the chain the user can set init_params to Vector{NamedTuple} whose elements are the starting locations for each of the nchains. If no starting location is specified nchains random samples from the prior will be chosen for the starting locations.

For possible kwargs please see the AdvancedHMC.jl docs

This returns a tuple where the first element is nchains of TypedTable's each which contains the MCMC samples of one of the parallel chain and the second argument is a set of ancilliary information about each set of samples.

Notes

This will automatically transform the posterior to the flattened unconstrained space.