ComradeOptimization

To optimize our posterior, we use the Optimization.jl package. Optimization provides a global interface to several Julia optimizers. The Comrade wrapper for Optimization.jl is very thin. The only difference addition is that Comrade has provided a method:

OptimizationFunction(::TransformedPosterior, args...; kwargs...)

meaning we can pass it a posterior object and it will set up the OptimizationFunction for us.

Note

We only specify this for a transformed version of the posterior. This is because Optimization.jl requires a flattened version of the posterior.

Additionally, different optimizers may prefer different parameter transformations. For example, if we use OptimizationBBO, using ascube is a good choice since it needs a compact region to search over, and ascube convert our parameter space to the unit hypercube. On the other hand, gradient-based optimizers work best without bounds, so a better choice would be the asflat transformation.

To see what optimizers are available and what options are available, please see the Optimizations.jl docs.

Example

using Comrade
using ComradeOptimization
using OptimizationOptimJL

# Some stuff to create a posterior object
post # of type Comrade.Posterior

# Create a optimization problem using Zygote as the backend
fflat = OptimizationProblem(asflat(post), Optimization.AutoZygote())

# create the problem from a random point in the prior, nothing is b/c there are no additional arugments to our function.
prob = OptimizationProblem(fflat, prior_sample(asflat(post)), nothing)

# Now solve! Here we use LBFGS
sol = solve(prob, LBFGS(); g_tol=1e-2)

API

ComradeOptimization.laplace — Method

laplace(prob, opt, args...; kwargs...)

Compute the Laplace or Quadratic approximation to the prob or posterior. The args and kwargs are passed the the SciMLBase.solve function. This will return a Distributions.MvNormal object that approximates the posterior in the transformed space.

Note the quadratic approximation is in the space of the transformed posterior not the usual parameter space. This is better for constrained problems where we may run up against a boundary.

SciMLBase.OptimizationFunction — Method

SciMLBase.OptimizationFunction(post::Posterior, args...; kwargs...)

Constructs a OptimizationFunction from a Comrade.TransformedPosterior object. Note that a user must transform the posterior first. This is so we know which space is most amenable to optimization.